@@ -2,18 +2,23 @@
  * Copyright (c) 2012-2016 Arne Schwabe
  * Distributed under the GNU GPL v2 with additional terms. For full terms see the file doc/LICENSE.txt
  */
+
 package de.blinkt.openvpn.core;
 
 import android.content.Context;
+import android.content.Intent;
 import android.net.LocalServerSocket;
 import android.net.LocalSocket;
 import android.net.LocalSocketAddress;
+import android.os.Build;
 import android.os.Handler;
 import android.os.ParcelFileDescriptor;
-import android.util.Log;
-import android.widget.Toast;
-
 import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
+import android.system.Os;
+import android.util.Log;
+import de.blinkt.openvpn.R;
+import de.blinkt.openvpn.VpnProfile;
 
 import java.io.FileDescriptor;
 import java.io.IOException;
@@ -22,18 +27,11 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.LinkedList;
-import java.util.Locale;
-import java.util.Vector;
-
-import de.blinkt.openvpn.BuildConfig;
-import de.blinkt.openvpn.R;
-import de.blinkt.openvpn.VpnProfile;
-import de.blinkt.openvpn.core.VpnStatus.ConnectionStatus;
+import java.util.*;
 
 public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
+
+    public static final int ORBOT_TIMEOUT_MS = 20 * 1000;
     private static final String TAG = "openvpn";
     private static final Vector<OpenVpnManagementThread> active = new Vector<>();
     private final Handler mResumeHandler;
@@ -45,22 +43,60 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
     private boolean mWaitingForRelease = false;
     private long mLastHoldRelease = 0;
     private LocalSocket mServerSocketLocal;
+
     private pauseReason lastPauseReason = pauseReason.noNetwork;
     private PausedStateCallback mPauseCallback;
     private boolean mShuttingDown;
-    private Runnable mResumeHoldRunnable = new Runnable() {
-        @Override
-        public void run() {
-            if (shouldBeRunning()) {
-                releaseHoldCmd();
-            }
+    private Runnable mResumeHoldRunnable = () -> {
+        if (shouldBeRunning()) {
+            releaseHoldCmd();
         }
     };
+    private Runnable orbotStatusTimeOutRunnable = new Runnable() {
+        @Override
+        public void run() {
+            sendProxyCMD(Connection.ProxyType.SOCKS5, "127.0.0.1", Integer.toString(OrbotHelper.SOCKS_PROXY_PORT_DEFAULT), false);
+            OrbotHelper.get(mOpenVPNService).removeStatusCallback(statusCallback);
+
+        }
+    };
+    private OrbotHelper.StatusCallback statusCallback = new OrbotHelper.StatusCallback() {
+
+        @Override
+        public void onStatus(Intent statusIntent) {
+            StringBuilder extras = new StringBuilder();
+            for (String key : statusIntent.getExtras().keySet()) {
+                Object val = statusIntent.getExtras().get(key);
+
+                extras.append(String.format(Locale.ENGLISH, "%s - '%s'", key, val == null ? "null" : val.toString()));
+            }
+            VpnStatus.logDebug("Got Orbot status: " + extras);
+        }
+
+        @Override
+        public void onNotYetInstalled() {
+            VpnStatus.logDebug("Orbot not yet installed");
+        }
+
+        @Override
+        public void onOrbotReady(Intent intent, String socksHost, int socksPort) {
+            mResumeHandler.removeCallbacks(orbotStatusTimeOutRunnable);
+            sendProxyCMD(Connection.ProxyType.SOCKS5, socksHost, Integer.toString(socksPort), false);
+            OrbotHelper.get(mOpenVPNService).removeStatusCallback(this);
+        }
+
+        @Override
+        public void onDisabled(Intent intent) {
+            VpnStatus.logWarning("Orbot integration for external applications is disabled. Waiting %ds before connecting to the default port. Enable external app integration in Orbot or use Socks v5 config instead of Orbot to avoid this delay.");
+        }
+    };
+    private transient Connection mCurrentProxyConnection;
 
     public OpenVpnManagementThread(VpnProfile profile, OpenVPNService openVpnService) {
         mProfile = profile;
         mOpenVPNService = openVpnService;
         mResumeHandler = new Handler(openVpnService.getMainLooper());
+
     }
 
     private static boolean stopOpenVPN() {
@@ -69,7 +105,8 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
             for (OpenVpnManagementThread mt : active) {
                 sendCMD = mt.managmentCommand("signal SIGINT\n");
                 try {
-                    if (mt.mSocket != null) mt.mSocket.close();
+                    if (mt.mSocket != null)
+                        mt.mSocket.close();
                 } catch (IOException e) {
                     // Ignore close error on already closed socket
                 }
@@ -81,28 +118,37 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
     public boolean openManagementInterface(@NonNull Context c) {
         // Could take a while to open connection
         int tries = 8;
+
         String socketName = (c.getCacheDir().getAbsolutePath() + "/" + "mgmtsocket");
         // The mServerSocketLocal is transferred to the LocalServerSocket, ignore warning
+
         mServerSocketLocal = new LocalSocket();
+
         while (tries > 0 && !mServerSocketLocal.isBound()) {
             try {
-                mServerSocketLocal.bind(new LocalSocketAddress(socketName, LocalSocketAddress.Namespace.FILESYSTEM));
+                mServerSocketLocal.bind(new LocalSocketAddress(socketName,
+                        LocalSocketAddress.Namespace.FILESYSTEM));
             } catch (IOException e) {
                 // wait 300 ms before retrying
                 try {
                     Thread.sleep(300);
                 } catch (InterruptedException ignored) {
                 }
+
             }
             tries--;
         }
+
         try {
+
             mServerSocket = new LocalServerSocket(mServerSocketLocal.getFileDescriptor());
             return true;
         } catch (IOException e) {
             VpnStatus.logException(e);
         }
         return false;
+
+
     }
 
     /**
@@ -126,25 +172,35 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
     public void run() {
         byte[] buffer = new byte[2048];
         //	mSocket.setSoTimeout(5); // Setting a timeout cannot be that bad
+
         String pendingInput = "";
         synchronized (active) {
             active.add(this);
         }
+
         try {
             // Wait for a client to connect
             mSocket = mServerSocket.accept();
             InputStream instream = mSocket.getInputStream();
+
+
             // Close the management socket after client connected
             try {
                 mServerSocket.close();
             } catch (IOException e) {
                 VpnStatus.logException(e);
             }
+
             // Closing one of the two sockets also closes the other
             //mServerSocketLocal.close();
+            managmentCommand("version 3\n");
+
             while (true) {
+
                 int numbytesread = instream.read(buffer);
-                if (numbytesread == -1) return;
+                if (numbytesread == -1)
+                    return;
+
                 FileDescriptor[] fds = null;
                 try {
                     fds = mSocket.getAncillaryFileDescriptors();
@@ -154,12 +210,18 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
                 if (fds != null) {
                     Collections.addAll(mFDList, fds);
                 }
+
                 String input = new String(buffer, 0, numbytesread, "UTF-8");
+
                 pendingInput += input;
+
                 pendingInput = processInput(pendingInput);
+
+
             }
         } catch (IOException e) {
-            if (!e.getMessage().equals("socket closed") && !e.getMessage().equals("Connection reset by peer")) VpnStatus.logException(e);
+            if (!e.getMessage().equals("socket closed") && !e.getMessage().equals("Connection reset by peer"))
+                VpnStatus.logException(e);
         }
         synchronized (active) {
             active.remove(this);
@@ -171,37 +233,62 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
         try {
             Method getInt = FileDescriptor.class.getDeclaredMethod("getInt$");
             int fdint = (Integer) getInt.invoke(fd);
+
             // You can even get more evil by parsing toString() and extract the int from that :)
+
             boolean result = mOpenVPNService.protect(fdint);
-            if (!result) VpnStatus.logWarning("Could not protect VPN socket");
+            if (!result)
+                VpnStatus.logWarning("Could not protect VPN socket");
+
+
             //ParcelFileDescriptor pfd = ParcelFileDescriptor.fromFd(fdint);
             //pfd.close();
-            NativeUtils.jniclose(fdint);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                fdCloseLollipop(fd);
+            } else {
+                NativeUtils.jniclose(fdint);
+            }
             return;
-        } catch (NoSuchMethodException | IllegalArgumentException | InvocationTargetException | IllegalAccessException | NullPointerException e) {
+        } catch ( NoSuchMethodException | IllegalArgumentException | InvocationTargetException | IllegalAccessException | NullPointerException e) {
             VpnStatus.logException("Failed to retrieve fd from socket (" + fd + ")", e);
         }
+
         Log.d("Openvpn", "Failed to retrieve fd from socket: " + fd);
+
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    private void fdCloseLollipop(FileDescriptor fd) {
+        try {
+            Os.close(fd);
+        } catch (Exception e) {
+            VpnStatus.logException("Failed to close fd (" + fd + ")", e);
+        }
     }
 
     private String processInput(String pendingInput) {
+
+
         while (pendingInput.contains("\n")) {
             String[] tokens = pendingInput.split("\\r?\\n", 2);
             processCommand(tokens[0]);
             if (tokens.length == 1)
                 // No second part, newline was at the end
                 pendingInput = "";
-            else pendingInput = tokens[1];
+            else
+                pendingInput = tokens[1];
         }
         return pendingInput;
     }
 
     private void processCommand(String command) {
         //Log.i(TAG, "Line from managment" + command);
+
         if (command.startsWith(">") && command.contains(":")) {
             String[] parts = command.split(":", 2);
             String cmd = parts[0].substring(1);
             String argument = parts[1];
+
             switch (cmd) {
                 case "INFO":
                 /* Ignore greeting from management */
@@ -219,7 +306,8 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
                     processByteCount(argument);
                     break;
                 case "STATE":
-                    if (!mShuttingDown) processState(argument);
+                    if (!mShuttingDown)
+                        processState(argument);
                     break;
                 case "PROXY":
                     processProxyCMD(argument);
@@ -227,8 +315,11 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
                 case "LOG":
                     processLogMessage(argument);
                     break;
-                case "RSA_SIGN":
+                case "PK_SIGN":
                     processSignCommand(argument);
+                    break;
+                case "INFOMSG":
+                    processInfoMessage(argument);
                     break;
                 default:
                     VpnStatus.logWarning("MGMT: Got unrecognized command" + command);
@@ -240,10 +331,23 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
             return;
         } else if (command.startsWith("PROTECTFD: ")) {
             FileDescriptor fdtoprotect = mFDList.pollFirst();
-            if (fdtoprotect != null) protectFileDescriptor(fdtoprotect);
+            if (fdtoprotect != null)
+                protectFileDescriptor(fdtoprotect);
         } else {
             Log.i(TAG, "Got unrecognized line from managment" + command);
             VpnStatus.logWarning("MGMT: Got unrecognized line from management:" + command);
+        }
+    }
+
+    private void processInfoMessage(String info)
+    {
+        if (info.startsWith("OPEN_URL:") || info.startsWith("CR_TEXT:"))
+        {
+            mOpenVPNService.trigger_sso(info);
+        }
+        else
+        {
+            VpnStatus.logDebug("Info message from server:" + info);
         }
     }
 
@@ -260,6 +364,9 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
           D -- debug, and
                  */
         // 2 log message
+
+        Log.d("OpenVPN", argument);
+
         VpnStatus.LogLevel level;
         switch (args[1]) {
             case "I":
@@ -278,27 +385,37 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
                 level = VpnStatus.LogLevel.INFO;
                 break;
         }
+
         int ovpnlevel = Integer.parseInt(args[2]) & 0x0F;
         String msg = args[3];
-        if (msg.startsWith("MANAGEMENT: CMD")) ovpnlevel = Math.max(4, ovpnlevel);
+
+        if (msg.startsWith("MANAGEMENT: CMD"))
+            ovpnlevel = Math.max(4, ovpnlevel);
+
         VpnStatus.logMessageOpenVPN(level, ovpnlevel, msg);
     }
 
     boolean shouldBeRunning() {
-        if (mPauseCallback == null) return false;
-        else return mPauseCallback.shouldBeRunning();
+        if (mPauseCallback == null)
+            return false;
+        else
+            return mPauseCallback.shouldBeRunning();
     }
 
     private void handleHold(String argument) {
+        mWaitingForRelease = true;
         int waittime = Integer.parseInt(argument.split(":")[1]);
         if (shouldBeRunning()) {
-            if (waittime > 1) VpnStatus.updateStateString("CONNECTRETRY", String.valueOf(waittime), R.string.state_waitconnectretry, ConnectionStatus.LEVEL_CONNECTING_NO_SERVER_REPLY_YET);
+            if (waittime > 1)
+                VpnStatus.updateStateString("CONNECTRETRY", String.valueOf(waittime),
+                        R.string.state_waitconnectretry, ConnectionStatus.LEVEL_CONNECTING_NO_SERVER_REPLY_YET);
             mResumeHandler.postDelayed(mResumeHoldRunnable, waittime * 1000);
-            if (waittime > 5) {
+            if (waittime > 5)
                 VpnStatus.logInfo(R.string.state_waitconnectretry, String.valueOf(waittime));
-            }
+            else
+                VpnStatus.logDebug(R.string.state_waitconnectretry, String.valueOf(waittime));
+
         } else {
-            mWaitingForRelease = true;
             VpnStatus.updateStatePause(lastPauseReason);
         }
     }
@@ -310,6 +427,7 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
                 Thread.sleep(3000);
             } catch (InterruptedException ignored) {
             }
+
         }
         mWaitingForRelease = false;
         mLastHoldRelease = System.currentTimeMillis();
@@ -320,22 +438,83 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
     }
 
     public void releaseHold() {
-        if (mWaitingForRelease) releaseHoldCmd();
+        if (mWaitingForRelease)
+            releaseHoldCmd();
     }
 
     private void processProxyCMD(String argument) {
         String[] args = argument.split(",", 3);
-        SocketAddress proxyaddr = ProxyDetection.detectProxy(mProfile);
-        if (args.length >= 2) {
-            String proto = args[1];
-            if (proto.equals("UDP")) {
-                proxyaddr = null;
+
+        Connection.ProxyType proxyType = Connection.ProxyType.NONE;
+
+        int connectionEntryNumber = Integer.parseInt(args[0]) - 1;
+        String proxyport = null;
+        String proxyname = null;
+        boolean proxyUseAuth = false;
+
+        if (mProfile.mConnections.length > connectionEntryNumber) {
+            Connection connection = mProfile.mConnections[connectionEntryNumber];
+            proxyType = connection.mProxyType;
+            proxyname = connection.mProxyName;
+            proxyport = connection.mProxyPort;
+            proxyUseAuth = connection.mUseProxyAuth;
+
+            // Use transient variable to remember http user/password
+            mCurrentProxyConnection = connection;
+
+        } else {
+            VpnStatus.logError(String.format(Locale.ENGLISH, "OpenVPN is asking for a proxy of an unknown connection entry (%d)", connectionEntryNumber));
+        }
+
+        // atuo detection of proxy
+        if (proxyType == Connection.ProxyType.NONE) {
+            SocketAddress proxyaddr = ProxyDetection.detectProxy(mProfile);
+            if (proxyaddr instanceof InetSocketAddress) {
+                InetSocketAddress isa = (InetSocketAddress) proxyaddr;
+                proxyType = Connection.ProxyType.HTTP;
+                proxyname = isa.getHostName();
+                proxyport = String.valueOf(isa.getPort());
+                proxyUseAuth = false;
+
             }
         }
-        if (proxyaddr instanceof InetSocketAddress) {
-            InetSocketAddress isa = (InetSocketAddress) proxyaddr;
-            VpnStatus.logInfo(R.string.using_proxy, isa.getHostName(), isa.getPort());
-            String proxycmd = String.format(Locale.ENGLISH, "proxy HTTP %s %d\n", isa.getHostName(), isa.getPort());
+
+
+        if (args.length >= 2 && proxyType == Connection.ProxyType.HTTP) {
+            String proto = args[1];
+            if (proto.equals("UDP")) {
+                proxyname = null;
+                VpnStatus.logInfo("Not using an HTTP proxy since the connection uses UDP");
+            }
+        }
+
+
+        if (proxyType == Connection.ProxyType.ORBOT) {
+            VpnStatus.updateStateString("WAIT_ORBOT", "Waiting for Orbot to start", R.string.state_waitorbot, ConnectionStatus.LEVEL_CONNECTING_NO_SERVER_REPLY_YET);
+            OrbotHelper orbotHelper = OrbotHelper.get(mOpenVPNService);
+            if (!orbotHelper.checkTorReceier(mOpenVPNService))
+                VpnStatus.logError("Orbot does not seem to be installed!");
+
+            mResumeHandler.postDelayed(orbotStatusTimeOutRunnable, ORBOT_TIMEOUT_MS);
+            orbotHelper.addStatusCallback(mOpenVPNService, statusCallback);
+
+            orbotHelper.sendOrbotStartAndStatusBroadcast();
+
+        } else {
+            sendProxyCMD(proxyType, proxyname, proxyport, proxyUseAuth);
+        }
+    }
+
+    private void sendProxyCMD(Connection.ProxyType proxyType, String proxyname, String proxyport, boolean usePwAuth) {
+        if (proxyType != Connection.ProxyType.NONE && proxyname != null) {
+
+            VpnStatus.logInfo(R.string.using_proxy, proxyname, proxyname);
+
+            String pwstr =  usePwAuth ? " auto" : "";
+
+            String proxycmd = String.format(Locale.ENGLISH, "proxy %s %s %s%s\n",
+                    proxyType == Connection.ProxyType.HTTP ? "HTTP" : "SOCKS",
+                    proxyname, proxyport, pwstr);
             managmentCommand(proxycmd);
         } else {
             managmentCommand("proxy NONE\n");
@@ -345,8 +524,11 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
     private void processState(String argument) {
         String[] args = argument.split(",", 3);
         String currentstate = args[1];
-        if (args[2].equals(",,")) VpnStatus.updateStateString(currentstate, "");
-        else VpnStatus.updateStateString(currentstate, args[2]);
+
+        if (args[2].equals(",,"))
+            VpnStatus.updateStateString(currentstate, "");
+        else
+            VpnStatus.updateStateString(currentstate, args[2]);
     }
 
     private void processByteCount(String argument) {
@@ -354,21 +536,28 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
         int comma = argument.indexOf(',');
         long in = Long.parseLong(argument.substring(0, comma));
         long out = Long.parseLong(argument.substring(comma + 1));
+
         VpnStatus.updateByteCount(in, out);
+
     }
 
     private void processNeedCommand(String argument) {
         int p1 = argument.indexOf('\'');
         int p2 = argument.indexOf('\'', p1 + 1);
+
         String needed = argument.substring(p1 + 1, p2);
         String extra = argument.split(":", 2)[1];
+
         String status = "ok";
+
+
         switch (needed) {
             case "PROTECTFD":
                 FileDescriptor fdtoprotect = mFDList.pollFirst();
                 protectFileDescriptor(fdtoprotect);
                 break;
             case "DNSSERVER":
+            case "DNS6SERVER":
                 mOpenVPNService.addDNS(extra);
                 break;
             case "DNSDOMAIN":
@@ -376,21 +565,23 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
                 break;
             case "ROUTE": {
                 String[] routeparts = extra.split(" ");
+
             /*
             buf_printf (&out, "%s %s %s dev %s", network, netmask, gateway, rgi->iface);
             else
             buf_printf (&out, "%s %s %s", network, netmask, gateway);
             */
+
                 if (routeparts.length == 5) {
-                    if (BuildConfig.DEBUG) {
-                        Log.i(TAG, "processNeedCommand: "+ routeparts[3]);
-                    }
+                    //if (BuildConfig.DEBUG)
+                    //                assertEquals("dev", routeparts[3]);
                     mOpenVPNService.addRoute(routeparts[0], routeparts[1], routeparts[2], routeparts[4]);
                 } else if (routeparts.length >= 3) {
                     mOpenVPNService.addRoute(routeparts[0], routeparts[1], routeparts[2], null);
                 } else {
                     VpnStatus.logError("Unrecognized ROUTE cmd:" + Arrays.toString(routeparts) + " | " + argument);
                 }
+
                 break;
             }
             case "ROUTE6": {
@@ -404,21 +595,28 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
                 mOpenVPNService.setLocalIP(ifconfigparts[0], ifconfigparts[1], mtu, ifconfigparts[3]);
                 break;
             case "IFCONFIG6":
-                mOpenVPNService.setLocalIPv6(extra);
+                String[] ifconfig6parts = extra.split(" ");
+                mtu = Integer.parseInt(ifconfig6parts[1]);
+                mOpenVPNService.setMtu(mtu);
+                mOpenVPNService.setLocalIPv6(ifconfig6parts[0]);
                 break;
             case "PERSIST_TUN_ACTION":
                 // check if tun cfg stayed the same
                 status = mOpenVPNService.getTunReopenStatus();
                 break;
             case "OPENTUN":
-                if (sendTunFD(needed, extra)) return;
-                else status = "cancel";
+                if (sendTunFD(needed, extra))
+                    return;
+                else
+                    status = "cancel";
                 // This not nice or anything but setFileDescriptors accepts only FilDescriptor class :(
+
                 break;
             default:
                 Log.e(TAG, "Unknown needok command " + argument);
                 return;
         }
+
         String cmd = String.format("needok '%s' %s\n", needed, status);
         managmentCommand(cmd);
     }
@@ -427,31 +625,41 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
         if (!extra.equals("tun")) {
             // We only support tun
             VpnStatus.logError(String.format("Device type %s requested, but only tun is possible with the Android API, sorry!", extra));
+
             return false;
         }
         ParcelFileDescriptor pfd = mOpenVPNService.openTun();
-        if (pfd == null) return false;
+        if (pfd == null)
+            return false;
+
         Method setInt;
         int fdint = pfd.getFd();
         try {
             setInt = FileDescriptor.class.getDeclaredMethod("setInt$", int.class);
             FileDescriptor fdtosend = new FileDescriptor();
+
             setInt.invoke(fdtosend, fdint);
+
             FileDescriptor[] fds = {fdtosend};
             mSocket.setFileDescriptorsForSend(fds);
+
             // Trigger a send so we can close the fd on our side of the channel
             // The API documentation fails to mention that it will not reset the file descriptor to
             // be send and will happily send the file descriptor on every write ...
             String cmd = String.format("needok '%s' %s\n", needed, "ok");
             managmentCommand(cmd);
+
             // Set the FileDescriptor to null to stop this mad behavior
             mSocket.setFileDescriptorsForSend(null);
+
             pfd.close();
+
             return true;
         } catch (NoSuchMethodException | IllegalArgumentException | InvocationTargetException |
                 IOException | IllegalAccessException exp) {
             VpnStatus.logException("Could not send fd over socket", exp);
         }
+
         return false;
     }
 
@@ -459,7 +667,14 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
         //argument has the form 	Need 'Private Key' password
         // or  ">PASSWORD:Verification Failed: '%s' ['%s']"
         String needed;
+
+
         try {
+            // Ignore Auth token message, already managed by openvpn itself
+            if (argument.startsWith("Auth-Token:")) {
+                return;
+            }
+
             int p1 = argument.indexOf('\'');
             int p2 = argument.indexOf('\'', p1 + 1);
             needed = argument.substring(p1 + 1, p2);
@@ -471,33 +686,49 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
             VpnStatus.logError("Could not parse management Password command: " + argument);
             return;
         }
+
         String pw = null;
+        String username = null;
+
         if (needed.equals("Private Key")) {
             pw = mProfile.getPasswordPrivateKey();
         } else if (needed.equals("Auth")) {
-            String usercmd = String.format("username '%s' %s\n", needed, VpnProfile.openVpnEscape(mProfile.mUsername));
-            managmentCommand(usercmd);
             pw = mProfile.getPasswordAuth();
+            username = mProfile.mUsername;
+
+        } else if (needed.equals("HTTP Proxy")) {
+            if( mCurrentProxyConnection != null) {
+                pw = mCurrentProxyConnection.mProxyAuthPassword;
+                username = mCurrentProxyConnection.mProxyAuthUser;
+            }
         }
         if (pw != null) {
+            if (username !=null) {
+                String usercmd = String.format("username '%s' %s\n",
+                        needed, VpnProfile.openVpnEscape(username));
+                managmentCommand(usercmd);
+            }
             String cmd = String.format("password '%s' %s\n", needed, VpnProfile.openVpnEscape(pw));
             managmentCommand(cmd);
         } else {
+            mOpenVPNService.requestInputFromUser(R.string.password, needed);
             VpnStatus.logError(String.format("Openvpn requires Authentication type '%s' but no password/key information available", needed));
         }
+
     }
 
-    static boolean  flag = false;
     private void proccessPWFailed(String needed, String args) {
-        flag = true;
         VpnStatus.updateStateString("AUTH_FAILED", needed + args, R.string.state_auth_failed, ConnectionStatus.LEVEL_AUTH_FAILED);
     }
 
     @Override
     public void networkChange(boolean samenetwork) {
-        if (mWaitingForRelease) releaseHold();
-        else if (samenetwork) managmentCommand("network-change samenetwork\n");
-        else managmentCommand("network-change\n");
+        if (mWaitingForRelease)
+            releaseHold();
+        else if (samenetwork)
+            managmentCommand("network-change samenetwork\n");
+        else
+            managmentCommand("network-change\n");
     }
 
     @Override
@@ -505,9 +736,15 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
         mPauseCallback = callback;
     }
 
+    @Override
+    public void sendCRResponse(String response) {
+        managmentCommand("cr-response "  + response + "\n");
+    }
+
     public void signalusr1() {
         mResumeHandler.removeCallbacks(mResumeHoldRunnable);
-        if (!mWaitingForRelease) managmentCommand("signal SIGUSR1\n");
+        if (!mWaitingForRelease)
+            managmentCommand("signal SIGUSR1\n");
         else
             // If signalusr1 is called update the state string
             // if there is another for stopping
@@ -519,15 +756,20 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
         releaseHold();
     }
 
-    private void processSignCommand(String b64data) {
-        String signed_string = mProfile.getSignedData(b64data);
+    private void processSignCommand(String argument) {
+
+        String[] arguments = argument.split(",");
+
+        boolean pkcs1padding = arguments[1].equals("RSA_PKCS1_PADDING");
+        String signed_string = mProfile.getSignedData(mOpenVPNService, arguments[0], pkcs1padding);
+
         if (signed_string == null) {
-            managmentCommand("rsa-sig\n");
+            managmentCommand("pk-sig\n");
             managmentCommand("\nEND\n");
             stopOpenVPN();
             return;
         }
-        managmentCommand("rsa-sig\n");
+        managmentCommand("pk-sig\n");
         managmentCommand(signed_string);
         managmentCommand("\nEND\n");
     }
@@ -550,7 +792,9 @@ public class OpenVpnManagementThread implements Runnable, OpenVPNManagement {
         boolean stopSucceed = stopOpenVPN();
         if (stopSucceed) {
             mShuttingDown = true;
+
         }
         return stopSucceed;
     }
+
 }
