@@ -11,12 +11,10 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.ImageButton;
 
 import com.lazycoder.cakevpn.R;
-import com.lazycoder.cakevpn.SharedPreference;
 import com.lazycoder.cakevpn.adapter.ServerListRVAdapter;
 import com.lazycoder.cakevpn.interfaces.ChangeServer;
 import com.lazycoder.cakevpn.interfaces.NavItemClickListener;
@@ -24,7 +22,7 @@ import com.lazycoder.cakevpn.model.Server;
 
 import java.util.ArrayList;
 
-import static com.lazycoder.cakevpn.Utils.getImgURL;
+import com.lazycoder.cakevpn.Utils;
 
 
 public class MainActivity extends AppCompatActivity implements NavItemClickListener {
@@ -111,25 +109,25 @@ public class MainActivity extends AppCompatActivity implements NavItemClickListe
         ArrayList<Server> servers = new ArrayList<>();
 
         servers.add(new Server("United States",
-                getImgURL(R.drawable.usa_flag),
+                Utils.getImgURL(R.drawable.usa_flag),
                 "us.ovpn",
                 "freeopenvpn",
                 "416248023"
         ));
         servers.add(new Server("Japan",
-                getImgURL(R.drawable.japan),
+                Utils.getImgURL(R.drawable.japan),
                 "japan.ovpn",
                 "vpn",
                 "vpn"
         ));
         servers.add(new Server("Sweden",
-                getImgURL(R.drawable.sweden),
+                Utils.getImgURL(R.drawable.sweden),
                 "sweden.ovpn",
                 "vpn",
                 "vpn"
         ));
         servers.add(new Server("Korea",
-                getImgURL(R.drawable.korea),
+                Utils.getImgURL(R.drawable.korea),
                 "korea.ovpn",
                 "vpn",
                 "vpn"
