@@ -20,26 +20,26 @@ public class OpenVpnApi {
 
     private static final String  TAG = "OpenVpnApi";
     @TargetApi(Build.VERSION_CODES.ICE_CREAM_SANDWICH_MR1)
-    public static void startVpn(Context context, String inlineConfig, String userName, String pw) throws RemoteException {
+    public static void startVpn(Context context, String inlineConfig, String sCountry, String userName, String pw) throws RemoteException {
         if (TextUtils.isEmpty(inlineConfig)) throw new RemoteException("config is empty");
-            startVpnInternal(context, inlineConfig, userName, pw);
+            startVpnInternal(context, inlineConfig, sCountry, userName, pw);
     }
 
-    static void startVpnInternal(Context context, String inlineConfig, String userName, String pw) throws RemoteException {
+    static void startVpnInternal(Context context, String inlineConfig, String sCountry, String userName, String pw) throws RemoteException {
         ConfigParser cp = new ConfigParser();
         try {
             cp.parseConfig(new StringReader(inlineConfig));
             VpnProfile vp = cp.convertProfile();// Analysis.ovpn
             Log.d(TAG, "startVpnInternal: =============="+cp+"\n" +
                     vp);
-            vp.mName = "VPN";
+            vp.mName = sCountry;
             if (vp.checkProfile(context) != de.blinkt.openvpn.R.string.no_error_found){
                 throw new RemoteException(context.getString(vp.checkProfile(context)));
             }
             vp.mProfileCreator = context.getPackageName();
             vp.mUsername = userName;
             vp.mPassword = pw;
-            ProfileManager.setTemporaryProfile(vp);
+            ProfileManager.setTemporaryProfile(context, vp);
             VPNLaunchHelper.startOpenVpn(vp, context);
         } catch (IOException | ConfigParser.ConfigParseError e) {
             throw new RemoteException(e.getMessage());
