@@ -2,6 +2,7 @@
  * Copyright (c) 2012-2016 Arne Schwabe
  * Distributed under the GNU GPL v2 with additional terms. For full terms see the file doc/LICENSE.txt
  */
+
 package de.blinkt.openvpn;
 
 import android.annotation.SuppressLint;
@@ -14,12 +15,12 @@ import android.os.Build;
 import android.preference.PreferenceManager;
 import android.security.KeyChain;
 import android.security.KeyChainException;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.text.TextUtils;
 import android.util.Base64;
-import android.util.Log;
 
-import androidx.annotation.NonNull;
-
+import de.blinkt.openvpn.core.*;
 import org.spongycastle.util.io.pem.PemObject;
 import org.spongycastle.util.io.pem.PemWriter;
 
@@ -32,9 +33,7 @@ import java.io.Serializable;
 import java.io.StringWriter;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
-import java.security.PrivateKey;
+import java.security.*;
 import java.security.cert.Certificate;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
@@ -49,40 +48,49 @@ import javax.crypto.Cipher;
 import javax.crypto.IllegalBlockSizeException;
 import javax.crypto.NoSuchPaddingException;
 
-import de.blinkt.openvpn.core.Connection;
-import de.blinkt.openvpn.core.NativeUtils;
-import de.blinkt.openvpn.core.OpenVPNService;
-import de.blinkt.openvpn.core.VPNLaunchHelper;
-import de.blinkt.openvpn.core.VpnStatus;
-import de.blinkt.openvpn.core.X509Utils;
-
 public class VpnProfile implements Serializable, Cloneable {
+    // Note that this class cannot be moved to core where it belongs since
+    // the profile loading depends on it being here
+    // The Serializable documentation mentions that class name change are possible
+    // but the how is unclear
+    //
+    transient public static final long MAX_EMBED_FILE_SIZE = 2048 * 1024; // 2048kB
+    // Don't change this, not all parts of the program use this constant
+    public static final String EXTRA_PROFILEUUID = "de.blinkt.openvpn.profileUUID";
     public static final String INLINE_TAG = "[[INLINE]]";
-    private static final String DISPLAYNAME_TAG = "[[NAME]]";
-    private static final int MAXLOGLEVEL = 4;
-    private static final int CURRENT_PROFILE_VERSION = 6;
+    public static final String DISPLAYNAME_TAG = "[[NAME]]";
+    public static final int MAXLOGLEVEL = 4;
+    public static final int CURRENT_PROFILE_VERSION = 8;
+    public static final int DEFAULT_MSSFIX_SIZE = 1280;
     public static final int TYPE_CERTIFICATES = 0;
-    private static final int TYPE_PKCS12 = 1;
+    public static final int TYPE_PKCS12 = 1;
     public static final int TYPE_KEYSTORE = 2;
     public static final int TYPE_USERPASS = 3;
     public static final int TYPE_STATICKEYS = 4;
     public static final int TYPE_USERPASS_CERTIFICATES = 5;
-    private static final int TYPE_USERPASS_PKCS12 = 6;
+    public static final int TYPE_USERPASS_PKCS12 = 6;
     public static final int TYPE_USERPASS_KEYSTORE = 7;
+    public static final int TYPE_EXTERNAL_APP = 8;
     public static final int X509_VERIFY_TLSREMOTE = 0;
     public static final int X509_VERIFY_TLSREMOTE_COMPAT_NOREMAPPING = 1;
     public static final int X509_VERIFY_TLSREMOTE_DN = 2;
     public static final int X509_VERIFY_TLSREMOTE_RDN = 3;
     public static final int X509_VERIFY_TLSREMOTE_RDN_PREFIX = 4;
-    private static final boolean mIsOpenVPN22 = false;
+    public static final int AUTH_RETRY_NONE_FORGET = 0;
+    public static final int AUTH_RETRY_NOINTERACT = 2;
+    public static final boolean mIsOpenVPN22 = false;
     private static final long serialVersionUID = 7085688938959334563L;
+    private static final int AUTH_RETRY_NONE_KEEP = 1;
+    private static final int AUTH_RETRY_INTERACT = 3;
     public static String DEFAULT_DNS1 = "8.8.8.8";
-    private static String DEFAULT_DNS2 = "8.8.4.4";
-    private transient String mTransientPW = null;
-    private transient String mTransientPCKS12PW = null;
+    public static String DEFAULT_DNS2 = "8.8.4.4";
+    // variable named wrong and should haven beeen transient
+    // but needs to keep wrong name to guarante loading of old
+    // profiles
+    public transient boolean profileDeleted = false;
     public int mAuthenticationType = TYPE_KEYSTORE;
     public String mName;
-    private String mAlias;
+    public String mAlias;
     public String mClientCertFilename;
     public String mTLSAuthDirection = "";
     public String mTLSAuthFilename;
@@ -90,16 +98,13 @@ public class VpnProfile implements Serializable, Cloneable {
     public String mCaFilename;
     public boolean mUseLzo = true;
     public String mPKCS12Filename;
-    private String mPKCS12Password;
+    public String mPKCS12Password;
     public boolean mUseTLSAuth = false;
     public String mDNS1 = DEFAULT_DNS1;
     public String mDNS2 = DEFAULT_DNS2;
     public String mIPv4Address;
-    private String mIPv6Address;
-    // issue #27 dns-leak solution
-    public boolean mOverrideDNS = true;
-    // public boolean mOverrideDNS = false;
-    
+    public String mIPv6Address;
+    public boolean mOverrideDNS = false;
     public String mSearchDomain = "blinkt.de";
     public boolean mUseDefaultRoute = true;
     public boolean mUsePull = true;
@@ -116,53 +121,78 @@ public class VpnProfile implements Serializable, Cloneable {
     public String mCustomConfigOptions = "";
     public String mVerb = "1";  //ignored
     public String mCipher = "";
-    public boolean mNobind = false;
-    private boolean mUseDefaultRoutev6 = true;
+    public boolean mNobind = true;
+    public boolean mUseDefaultRoutev6 = true;
     public String mCustomRoutesv6 = "";
-    private String mKeyPassword = "";
+    public String mKeyPassword = "";
     public boolean mPersistTun = false;
     public String mConnectRetryMax = "-1";
     public String mConnectRetry = "2";
     public String mConnectRetryMaxTime = "300";
-    private boolean mUserEditable = true;
+    public boolean mUserEditable = true;
     public String mAuth = "";
     public int mX509AuthType = X509_VERIFY_TLSREMOTE_RDN;
     public String mx509UsernameField = null;
     public boolean mAllowLocalLAN;
     public String mExcludedRoutes;
+    public String mExcludedRoutesv6;
     public int mMssFix = 0; // -1 is default,
     public Connection[] mConnections = new Connection[0];
     public boolean mRemoteRandom = false;
     public HashSet<String> mAllowedAppsVpn = new HashSet<>();
     public boolean mAllowedAppsVpnAreDisallowed = true;
+    public boolean mAllowAppVpnBypass = false;
     public String mCrlFilename;
-    String mProfileCreator;
+    public String mProfileCreator;
+    public String mExternalAuthenticator;
+    public int mAuthRetry = AUTH_RETRY_NONE_FORGET;
+    public int mTunMtu;
     public boolean mPushPeerInfo = false;
+    public int mVersion = 0;
+    // timestamp when the profile was last used
+    public long mLastUsed;
+    public String importedProfileHash;
     /* Options no longer used in new profiles */
-    public String mServerName = "openvpn.blinkt.de";
+    public String mServerName = "openvpn.example.com";
     public String mServerPort = "1194";
+    public boolean mUseUdp = true;
+    public boolean mTemporaryProfile = false;
     private transient PrivateKey mPrivateKey;
     // Public attributes, since I got mad with getter/setter
     // set members to default values
     private UUID mUuid;
     private int mProfileVersion;
 
+    public boolean mBlockUnusedAddressFamilies =true;
+
     public VpnProfile(String name) {
         mUuid = UUID.randomUUID();
         mName = name;
         mProfileVersion = CURRENT_PROFILE_VERSION;
+
         mConnections = new Connection[1];
         mConnections[0] = new Connection();
+        mLastUsed = System.currentTimeMillis();
     }
 
     public static String openVpnEscape(String unescaped) {
-        if (unescaped == null) return null;
+        if (unescaped == null)
+            return null;
         String escapedString = unescaped.replace("\\", "\\\\");
         escapedString = escapedString.replace("\"", "\\\"");
         escapedString = escapedString.replace("\n", "\\n");
+
         if (escapedString.equals(unescaped) && !escapedString.contains(" ") &&
-                !escapedString.contains("#") && !escapedString.contains(";") && !escapedString.equals("")) return unescaped;
-        else return '"' + escapedString + '"';
+                !escapedString.contains("#") && !escapedString.contains(";")
+                && !escapedString.equals(""))
+            return unescaped;
+        else
+            return '"' + escapedString + '"';
+    }
+
+    public static boolean doUseOpenVPN3(Context c) {
+        // Nerver use OpenVPN3
+        return false;
     }
 
     //! Put inline data inline and other data as normal escaped filename
@@ -184,13 +214,30 @@ public class VpnProfile implements Serializable, Cloneable {
     }
 
     public static String getEmbeddedContent(String data) {
-        if (!data.contains(INLINE_TAG)) return data;
+        if (!data.contains(INLINE_TAG))
+            return data;
+
         int start = data.indexOf(INLINE_TAG) + INLINE_TAG.length();
         return data.substring(start);
     }
 
     public static boolean isEmbedded(String data) {
-        return data != null && (data.startsWith(INLINE_TAG) || data.startsWith(DISPLAYNAME_TAG));
+        if (data == null)
+            return false;
+        if (data.startsWith(INLINE_TAG) || data.startsWith(DISPLAYNAME_TAG))
+            return true;
+        else
+            return false;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj instanceof VpnProfile) {
+            VpnProfile vpnProfile = (VpnProfile) obj;
+            return mUuid.equals(vpnProfile.mUuid);
+        } else {
+            return false;
+        }
     }
 
     public void clearDefaults() {
@@ -205,245 +252,425 @@ public class VpnProfile implements Serializable, Cloneable {
         mAllowLocalLAN = true;
         mPushPeerInfo = false;
         mMssFix = 0;
+        mNobind = false;
     }
 
     public UUID getUUID() {
         return mUuid;
+
+    }
+
+    // Only used for the special case of managed profiles
+    public void setUUID(UUID uuid) {
+        mUuid = uuid;
     }
 
     public String getName() {
-        if (mName == null) return "No profile name";
+        if (TextUtils.isEmpty(mName))
+            return "No profile name";
         return mName;
     }
 
     public void upgradeProfile() {
-        if (mProfileVersion < 2) {
-            /* default to the behaviour the OS used */
-            mAllowLocalLAN = Build.VERSION.SDK_INT < Build.VERSION_CODES.KITKAT;
+
+        /* Fallthrough is intended here */
+        switch(mProfileVersion) {
+            case 0:
+            case 1:
+                /* default to the behaviour the OS used */
+                mAllowLocalLAN = Build.VERSION.SDK_INT < Build.VERSION_CODES.KITKAT;
+            case 2:
+            case 3:
+                moveOptionsToConnection();
+                mAllowedAppsVpnAreDisallowed = true;
+
+                if (mAllowedAppsVpn == null)
+                    mAllowedAppsVpn = new HashSet<>();
+
+                if (mConnections == null)
+                    mConnections = new Connection[0];
+            case 4:
+            case 5:
+
+                if (TextUtils.isEmpty(mProfileCreator))
+                    mUserEditable = true;
+            case 6:
+                for (Connection c : mConnections)
+                    if (c.mProxyType == null)
+                        c.mProxyType = Connection.ProxyType.NONE;
+            case 7:
+                if (mAllowAppVpnBypass)
+                    mBlockUnusedAddressFamilies = !mAllowAppVpnBypass;
+            default:
         }
-        if (mProfileVersion < 4) {
-            moveOptionsToConnection();
-            mAllowedAppsVpnAreDisallowed = true;
-        }
-        if (mAllowedAppsVpn == null) mAllowedAppsVpn = new HashSet<>();
-        if (mConnections == null) mConnections = new Connection[0];
-        if (mProfileVersion < 6) {
-            if (TextUtils.isEmpty(mProfileCreator)) mUserEditable = true;
-        }
+
         mProfileVersion = CURRENT_PROFILE_VERSION;
+
     }
 
     private void moveOptionsToConnection() {
         mConnections = new Connection[1];
         Connection conn = new Connection();
+
         conn.mServerName = mServerName;
         conn.mServerPort = mServerPort;
-        conn.mUseUdp = true;
+        conn.mUseUdp = mUseUdp;
         conn.mCustomConfiguration = "";
+
         mConnections[0] = conn;
+
     }
 
-    private String getConfigFile(Context context, boolean configForOvpn3) {
+    public String getConfigFile(Context context, boolean configForOvpn3) {
+
         File cacheDir = context.getCacheDir();
-        String cfg = "";
-        // Enable management interface
-        cfg += "# Enables connection to GUI\n";
-        cfg += "management ";
-        cfg += cacheDir.getAbsolutePath() + "/" + "mgmtsocket";
-        cfg += " unix\n";
-        cfg += "management-client\n";
-        // Not needed, see updated man page in 2.3
-        //cfg += "management-signal\n";
-        cfg += "management-query-passwords\n";
-        cfg += "management-hold\n\n";
+        StringBuilder cfg = new StringBuilder();
+
         if (!configForOvpn3) {
-            cfg += String.format("setenv IV_GUI_VER %s \n", openVpnEscape(getVersionEnvString(context)));
-            String versionString = String.format(Locale.US, "%d %s %s %s %s %s", Build.VERSION.SDK_INT, Build.VERSION.RELEASE, NativeUtils.getNativeAPI(), Build.BRAND, Build.BOARD, Build.MODEL);
-            cfg += String.format("setenv IV_PLAT_VER %s\n", openVpnEscape(versionString));
+            // Enable management interface
+            cfg.append("# Config for OpenVPN 2.x\n");
+            cfg.append("# Enables connection to GUI\n");
+            cfg.append("management ");
+
+            cfg.append(cacheDir.getAbsolutePath()).append("/").append("mgmtsocket");
+            cfg.append(" unix\n");
+            cfg.append("management-client\n");
+            // Not needed, see updated man page in 2.3
+            //cfg += "management-signal\n";
+            cfg.append("management-query-passwords\n");
+            cfg.append("management-hold\n\n");
+
+            cfg.append(String.format("setenv IV_GUI_VER %s \n", openVpnEscape(getVersionEnvString(context))));
+            cfg.append("setenv IV_SSO openurl,crtext\n");
+            String versionString = getPlatformVersionEnvString();
+            cfg.append(String.format("setenv IV_PLAT_VER %s\n", openVpnEscape(versionString)));
+        } else {
+            cfg.append("# Config for OpenVPN 3 C++\n");
         }
-        cfg += "machine-readable-output\n";
-        // Users are confused by warnings that are misleading...
-        cfg += "ifconfig-nowarn\n";
+
+
+        if (!configForOvpn3) {
+            cfg.append("machine-readable-output\n");
+            if (!mIsOpenVPN22)
+                cfg.append("allow-recursive-routing\n");
+
+            // Users are confused by warnings that are misleading...
+            cfg.append("ifconfig-nowarn\n");
+        }
+
         boolean useTLSClient = (mAuthenticationType != TYPE_STATICKEYS);
-        if (useTLSClient && mUsePull) cfg += "client\n";
-        else if (mUsePull) cfg += "pull\n";
-        else if (useTLSClient) cfg += "tls-client\n";
+
+        if (useTLSClient && mUsePull)
+            cfg.append("client\n");
+        else if (mUsePull)
+            cfg.append("pull\n");
+        else if (useTLSClient)
+            cfg.append("tls-client\n");
+
+
         //cfg += "verb " + mVerb + "\n";
-        cfg += "verb " + MAXLOGLEVEL + "\n";
+        cfg.append("verb " + MAXLOGLEVEL + "\n");
+
         if (mConnectRetryMax == null) {
             mConnectRetryMax = "-1";
         }
-        if (!mConnectRetryMax.equals("-1")) cfg += "connect-retry-max " + mConnectRetryMax + "\n";
-        if (TextUtils.isEmpty(mConnectRetry)) mConnectRetry = "2";
-        if (TextUtils.isEmpty(mConnectRetryMaxTime)) mConnectRetryMaxTime = "300";
-        boolean mUseUdp = true;
-        if (!mIsOpenVPN22) cfg += "connect-retry " + mConnectRetry + " " + mConnectRetryMaxTime + "\n";
-        else if (mIsOpenVPN22 && mUseUdp) cfg += "connect-retry " + mConnectRetry + "\n";
-        cfg += "resolv-retry 60\n";
+
+        if (!mConnectRetryMax.equals("-1"))
+            cfg.append("connect-retry-max ").append(mConnectRetryMax).append("\n");
+
+        if (TextUtils.isEmpty(mConnectRetry))
+            mConnectRetry = "2";
+
+        if (TextUtils.isEmpty(mConnectRetryMaxTime))
+            mConnectRetryMaxTime = "300";
+
+
+        if (!mIsOpenVPN22)
+            cfg.append("connect-retry ").append(mConnectRetry).append(" ").append(mConnectRetryMaxTime).append("\n");
+        else if (mIsOpenVPN22 && !mUseUdp)
+            cfg.append("connect-retry ").append(mConnectRetry).append("\n");
+
+
+        cfg.append("resolv-retry 60\n");
+
+
         // We cannot use anything else than tun
-        cfg += "dev tun\n";
+        cfg.append("dev tun\n");
+
+
         boolean canUsePlainRemotes = true;
+
         if (mConnections.length == 1) {
-            cfg += mConnections[0].getConnectionBlock();
+            cfg.append(mConnections[0].getConnectionBlock(configForOvpn3));
         } else {
             for (Connection conn : mConnections) {
                 canUsePlainRemotes = canUsePlainRemotes && conn.isOnlyRemote();
             }
-            if (mRemoteRandom) cfg += "remote-random\n";
+
+            if (mRemoteRandom)
+                cfg.append("remote-random\n");
+
             if (canUsePlainRemotes) {
                 for (Connection conn : mConnections) {
                     if (conn.mEnabled) {
-                        cfg += conn.getConnectionBlock();
+                        cfg.append(conn.getConnectionBlock(configForOvpn3));
                     }
                 }
             }
         }
+
+
         switch (mAuthenticationType) {
             case VpnProfile.TYPE_USERPASS_CERTIFICATES:
-                cfg += "auth-user-pass\n";
+                cfg.append("auth-user-pass\n");
             case VpnProfile.TYPE_CERTIFICATES:
                 // Ca
-                cfg += insertFileData("ca", mCaFilename);
+                cfg.append(insertFileData("ca", mCaFilename));
+
                 // Client Cert + Key
-                cfg += insertFileData("key", mClientKeyFilename);
-                cfg += insertFileData("cert", mClientCertFilename);
+                cfg.append(insertFileData("key", mClientKeyFilename));
+                cfg.append(insertFileData("cert", mClientCertFilename));
+
                 break;
             case VpnProfile.TYPE_USERPASS_PKCS12:
-                cfg += "auth-user-pass\n";
+                cfg.append("auth-user-pass\n");
             case VpnProfile.TYPE_PKCS12:
-                cfg += insertFileData("pkcs12", mPKCS12Filename);
+                cfg.append(insertFileData("pkcs12", mPKCS12Filename));
+
+                if (!TextUtils.isEmpty(mCaFilename))
+                {
+                    cfg.append(insertFileData("ca", mCaFilename));
+                }
                 break;
+
             case VpnProfile.TYPE_USERPASS_KEYSTORE:
-                cfg += "auth-user-pass\n";
+                cfg.append("auth-user-pass\n");
             case VpnProfile.TYPE_KEYSTORE:
+            case VpnProfile.TYPE_EXTERNAL_APP:
                 if (!configForOvpn3) {
-                    String[] ks = getKeyStoreCertificates(context);
-                    cfg += "### From Keystore ####\n";
+                    String[] ks = getExternalCertificates(context);
+                    cfg.append("### From Keystore/ext auth app ####\n");
                     if (ks != null) {
-                        cfg += "<ca>\n" + ks[0] + "\n</ca>\n";
-                        if (ks[1] != null) cfg += "<extra-certs>\n" + ks[1] + "\n</extra-certs>\n";
-                        cfg += "<cert>\n" + ks[2] + "\n</cert>\n";
-                        cfg += "management-external-key\n";
+                        cfg.append("<ca>\n").append(ks[0]).append("\n</ca>\n");
+                        if (!TextUtils.isEmpty(ks[1]))
+                            cfg.append("<extra-certs>\n").append(ks[1]).append("\n</extra-certs>\n");
+                        cfg.append("<cert>\n").append(ks[2]).append("\n</cert>\n");
+                        cfg.append("management-external-key nopadding\n");
                     } else {
-                        cfg += context.getString(R.string.keychain_access) + "\n";
-                        if (Build.VERSION.SDK_INT == Build.VERSION_CODES.JELLY_BEAN) if (!mAlias.matches("^[a-zA-Z0-9]$")) cfg += context.getString(R.string.jelly_keystore_alphanumeric_bug) + "\n";
+                        cfg.append(context.getString(R.string.keychain_access)).append("\n");
+                        if (Build.VERSION.SDK_INT == Build.VERSION_CODES.JELLY_BEAN)
+                            if (!mAlias.matches("^[a-zA-Z0-9]$"))
+                                cfg.append(context.getString(R.string.jelly_keystore_alphanumeric_bug)).append("\n");
                     }
                 }
                 break;
             case VpnProfile.TYPE_USERPASS:
-                cfg += "auth-user-pass\n";
-                cfg += insertFileData("ca", mCaFilename);
+                cfg.append("auth-user-pass\n");
+                cfg.append(insertFileData("ca", mCaFilename));
+                if (configForOvpn3) {
+                    // OpenVPN 3 needs to be told that a client certificate is not required
+                    cfg.append("client-cert-not-required\n");
+                }
         }
-        if (!TextUtils.isEmpty(mCrlFilename)) cfg += insertFileData("crl-verify", mCrlFilename);
+
+        if (isUserPWAuth()) {
+            if (mAuthRetry == AUTH_RETRY_NOINTERACT)
+                cfg.append("auth-retry nointeract\n");
+        }
+
+        if (!TextUtils.isEmpty(mCrlFilename))
+            cfg.append(insertFileData("crl-verify", mCrlFilename));
+
         if (mUseLzo) {
-            cfg += "comp-lzo\n";
+            cfg.append("comp-lzo\n");
         }
+
         if (mUseTLSAuth) {
-            if (mAuthenticationType == TYPE_STATICKEYS) cfg += insertFileData("secret", mTLSAuthFilename);
-            else cfg += insertFileData("tls-auth", mTLSAuthFilename);
-            if (!TextUtils.isEmpty(mTLSAuthDirection)) {
-                cfg += "key-direction ";
-                cfg += mTLSAuthDirection;
-                cfg += "\n";
+            boolean useTlsCrypt = mTLSAuthDirection.equals("tls-crypt");
+            boolean useTlsCrypt2 = mTLSAuthDirection.equals("tls-crypt-v2");
+
+            if (mAuthenticationType == TYPE_STATICKEYS)
+                cfg.append(insertFileData("secret", mTLSAuthFilename));
+            else if (useTlsCrypt)
+                cfg.append(insertFileData("tls-crypt", mTLSAuthFilename));
+            else if (useTlsCrypt2)
+                cfg.append(insertFileData("tls-crypt-v2", mTLSAuthFilename));
+            else
+                cfg.append(insertFileData("tls-auth", mTLSAuthFilename));
+
+            if (!TextUtils.isEmpty(mTLSAuthDirection) && !useTlsCrypt && !useTlsCrypt2) {
+                cfg.append("key-direction ");
+                cfg.append(mTLSAuthDirection);
+                cfg.append("\n");
             }
+
         }
+
         if (!mUsePull) {
-            if (!TextUtils.isEmpty(mIPv4Address)) cfg += "ifconfig " + cidrToIPAndNetmask(mIPv4Address) + "\n";
-            if (!TextUtils.isEmpty(mIPv6Address)) cfg += "ifconfig-ipv6 " + mIPv6Address + "\n";
+            if (!TextUtils.isEmpty(mIPv4Address))
+                cfg.append("ifconfig ").append(cidrToIPAndNetmask(mIPv4Address)).append("\n");
+
+            if (!TextUtils.isEmpty(mIPv6Address)) {
+                // Use our own ip as gateway since we ignore it anyway
+                String fakegw = mIPv6Address.split("/", 2)[0];
+                cfg.append("ifconfig-ipv6 ").append(mIPv6Address).append(" ").append(fakegw).append("\n");
+            }
+
         }
-        if (mUsePull && mRoutenopull) cfg += "route-nopull\n";
+
+        if (mUsePull && mRoutenopull)
+            cfg.append("route-nopull\n");
+
         String routes = "";
-        if (mUseDefaultRoute) routes += "route 0.0.0.0 0.0.0.0 vpn_gateway\n";
+
+        if (mUseDefaultRoute)
+            routes += "route 0.0.0.0 0.0.0.0 vpn_gateway\n";
         else {
             for (String route : getCustomRoutes(mCustomRoutes)) {
                 routes += "route " + route + " vpn_gateway\n";
             }
+
             for (String route : getCustomRoutes(mExcludedRoutes)) {
                 routes += "route " + route + " net_gateway\n";
             }
         }
-        if (mUseDefaultRoutev6) cfg += "route-ipv6 ::/0\n";
-        else for (String route : getCustomRoutesv6(mCustomRoutesv6)) {
-            routes += "route-ipv6 " + route + "\n";
-        }
-        cfg += routes;
+
+
+        if (mUseDefaultRoutev6)
+            cfg.append("route-ipv6 ::/0\n");
+        else
+            for (String route : getCustomRoutesv6(mCustomRoutesv6)) {
+                routes += "route-ipv6 " + route + "\n";
+            }
+
+        cfg.append(routes);
+
         if (mOverrideDNS || !mUsePull) {
-            if (!TextUtils.isEmpty(mDNS1)) cfg += "dhcp-option DNS " + mDNS1 + "\n";
-            if (!TextUtils.isEmpty(mDNS2)) cfg += "dhcp-option DNS " + mDNS2 + "\n";
-            if (!TextUtils.isEmpty(mSearchDomain)) cfg += "dhcp-option DOMAIN " + mSearchDomain + "\n";
+            if (!TextUtils.isEmpty(mDNS1)) {
+                cfg.append("dhcp-option DNS ").append(mDNS1).append("\n");
+            }
+            if (!TextUtils.isEmpty(mDNS2)) {
+                cfg.append("dhcp-option DNS ").append(mDNS2).append("\n");
+            }
+            if (!TextUtils.isEmpty(mSearchDomain))
+                cfg.append("dhcp-option DOMAIN ").append(mSearchDomain).append("\n");
+
         }
+
         if (mMssFix != 0) {
             if (mMssFix != 1450) {
-                cfg += String.format(Locale.US, "mssfix %d\n", mMssFix);
-            } else cfg += "mssfix\n";
+                if (configForOvpn3)
+                    cfg.append(String.format(Locale.US, "mssfix %d mtu\n", mMssFix));
+                else
+                    cfg.append(String.format(Locale.US, "mssfix %d\n", mMssFix));
+            } else
+                cfg.append("mssfix\n");
         }
-        if (mNobind) cfg += "nobind\n";
+
+        if (mTunMtu >= 48 && mTunMtu != 1500) {
+            cfg.append(String.format(Locale.US, "tun-mtu %d\n", mTunMtu));
+        }
+
+        if (mNobind)
+            cfg.append("nobind\n");
+
+
         // Authentication
         if (mAuthenticationType != TYPE_STATICKEYS) {
             if (mCheckRemoteCN) {
-                if (mRemoteCN == null || mRemoteCN.equals("")) cfg += "verify-x509-name " + openVpnEscape(mConnections[0].mServerName) + " name\n";
-                else switch (mX509AuthType) {
-                    // 2.2 style x509 checks
-                    case X509_VERIFY_TLSREMOTE_COMPAT_NOREMAPPING:
-                        cfg += "compat-names no-remapping\n";
-                    case X509_VERIFY_TLSREMOTE:
-                        cfg += "tls-remote " + openVpnEscape(mRemoteCN) + "\n";
-                        break;
-                    case X509_VERIFY_TLSREMOTE_RDN:
-                        cfg += "verify-x509-name " + openVpnEscape(mRemoteCN) + " name\n";
-                        break;
-                    case X509_VERIFY_TLSREMOTE_RDN_PREFIX:
-                        cfg += "verify-x509-name " + openVpnEscape(mRemoteCN) + " name-prefix\n";
-                        break;
-                    case X509_VERIFY_TLSREMOTE_DN:
-                        cfg += "verify-x509-name " + openVpnEscape(mRemoteCN) + "\n";
-                        break;
-                }
-                if (!TextUtils.isEmpty(mx509UsernameField)) cfg += "x509-username-field " + openVpnEscape(mx509UsernameField) + "\n";
+                if (mRemoteCN == null || mRemoteCN.equals(""))
+                    cfg.append("verify-x509-name ").append(openVpnEscape(mConnections[0].mServerName)).append(" name\n");
+                else
+                    switch (mX509AuthType) {
+
+                        // 2.2 style x509 checks
+                        case X509_VERIFY_TLSREMOTE_COMPAT_NOREMAPPING:
+                            cfg.append("compat-names no-remapping\n");
+                        case X509_VERIFY_TLSREMOTE:
+                            cfg.append("tls-remote ").append(openVpnEscape(mRemoteCN)).append("\n");
+                            break;
+
+                        case X509_VERIFY_TLSREMOTE_RDN:
+                            cfg.append("verify-x509-name ").append(openVpnEscape(mRemoteCN)).append(" name\n");
+                            break;
+
+                        case X509_VERIFY_TLSREMOTE_RDN_PREFIX:
+                            cfg.append("verify-x509-name ").append(openVpnEscape(mRemoteCN)).append(" name-prefix\n");
+                            break;
+
+                        case X509_VERIFY_TLSREMOTE_DN:
+                            cfg.append("verify-x509-name ").append(openVpnEscape(mRemoteCN)).append("\n");
+                            break;
+                    }
+                if (!TextUtils.isEmpty(mx509UsernameField))
+                    cfg.append("x509-username-field ").append(openVpnEscape(mx509UsernameField)).append("\n");
             }
-            if (mExpectTLSCert) cfg += "remote-cert-tls server\n";
+            if (mExpectTLSCert)
+                cfg.append("remote-cert-tls server\n");
         }
+
         if (!TextUtils.isEmpty(mCipher)) {
-            cfg += "cipher " + mCipher + "\n";
+            cfg.append("cipher ").append(mCipher).append("\n");
         }
+
         if (!TextUtils.isEmpty(mAuth)) {
-            cfg += "auth " + mAuth + "\n";
+            cfg.append("auth ").append(mAuth).append("\n");
         }
+
         // Obscure Settings dialog
-        if (mUseRandomHostname) cfg += "#my favorite options :)\nremote-random-hostname\n";
-        if (mUseFloat) cfg += "float\n";
+        if (mUseRandomHostname)
+            cfg.append("#my favorite options :)\nremote-random-hostname\n");
+
+        if (mUseFloat)
+            cfg.append("float\n");
+
         if (mPersistTun) {
-            cfg += "persist-tun\n";
-            cfg += "# persist-tun also enables pre resolving to avoid DNS resolve problem\n";
-            cfg += "preresolve\n";
+            cfg.append("persist-tun\n");
+            cfg.append("# persist-tun also enables pre resolving to avoid DNS resolve problem\n");
+            if (!mIsOpenVPN22)
+                cfg.append("preresolve\n");
         }
-        if (mPushPeerInfo) cfg += "push-peer-info\n";
+
+        if (mPushPeerInfo)
+            cfg.append("push-peer-info\n");
+
         SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
         boolean usesystemproxy = prefs.getBoolean("usesystemproxy", true);
-        if (usesystemproxy && !mIsOpenVPN22) {
-            cfg += "# Use system proxy setting\n";
-            cfg += "management-query-proxy\n";
+        if (usesystemproxy && !mIsOpenVPN22 && !configForOvpn3 && !usesExtraProxyOptions()) {
+            cfg.append("# Use system proxy setting\n");
+            cfg.append("management-query-proxy\n");
         }
+
+
         if (mUseCustomConfig) {
-            cfg += "# Custom configuration options\n";
-            cfg += "# You are on your on own here :)\n";
-            cfg += mCustomConfigOptions;
-            cfg += "\n";
+            cfg.append("# Custom configuration options\n");
+            cfg.append("# You are on your on own here :)\n");
+            cfg.append(mCustomConfigOptions);
+            cfg.append("\n");
+
         }
+
         if (!canUsePlainRemotes) {
-            cfg += "# Connection Options are at the end to allow global options (and global custom options) to influence connection blocks\n";
+            cfg.append("# Connection Options are at the end to allow global options (and global custom options) to influence connection blocks\n");
             for (Connection conn : mConnections) {
                 if (conn.mEnabled) {
-                    cfg += "<connection>\n";
-                    cfg += conn.getConnectionBlock();
-                    cfg += "</connection>\n";
+                    cfg.append("<connection>\n");
+                    cfg.append(conn.getConnectionBlock(configForOvpn3));
+                    cfg.append("</connection>\n");
                 }
             }
         }
-        return cfg;
+
+
+        return cfg.toString();
     }
 
-    private String getVersionEnvString(Context c) {
+    public String getPlatformVersionEnvString() {
+        return String.format(Locale.US, "%d %s %s %s %s %s", Build.VERSION.SDK_INT, Build.VERSION.RELEASE,
+                NativeUtils.getNativeAPI(), Build.BRAND, Build.BOARD, Build.MODEL);
+    }
+
+    static public String getVersionEnvString(Context c) {
         String version = "unknown";
         try {
             PackageInfo packageinfo = c.getPackageManager().getPackageInfo(c.getPackageName(), 0);
@@ -452,6 +679,7 @@ public class VpnProfile implements Serializable, Cloneable {
             VpnStatus.logException(e);
         }
         return String.format(Locale.US, "%s %s", c.getPackageName(), version);
+
     }
 
     @NonNull
@@ -464,10 +692,13 @@ public class VpnProfile implements Serializable, Cloneable {
         for (String route : routes.split("[\n \t]")) {
             if (!route.equals("")) {
                 String cidrroute = cidrToIPAndNetmask(route);
-                if (cidrroute == null) return cidrRoutes;
+                if (cidrroute == null)
+                    return cidrRoutes;
+
                 cidrRoutes.add(cidrroute);
             }
         }
+
         return cidrRoutes;
     }
 
@@ -482,29 +713,45 @@ public class VpnProfile implements Serializable, Cloneable {
                 cidrRoutes.add(route);
             }
         }
+
         return cidrRoutes;
     }
 
     private String cidrToIPAndNetmask(String route) {
         String[] parts = route.split("/");
+
         // No /xx, assume /32 as netmask
-        if (parts.length == 1) parts = (route + "/32").split("/");
-        if (parts.length != 2) return null;
+        if (parts.length == 1)
+            parts = (route + "/32").split("/");
+
+        if (parts.length != 2)
+            return null;
         int len;
         try {
             len = Integer.parseInt(parts[1]);
         } catch (NumberFormatException ne) {
             return null;
         }
-        if (len < 0 || len > 32) return null;
+        if (len < 0 || len > 32)
+            return null;
+
+
         long nm = 0xffffffffL;
         nm = (nm << (32 - len)) & 0xffffffffL;
+
         String netmask = String.format(Locale.ENGLISH, "%d.%d.%d.%d", (nm & 0xff000000) >> 24, (nm & 0xff0000) >> 16, (nm & 0xff00) >> 8, nm & 0xff);
         return parts[0] + "  " + netmask;
     }
 
     public Intent prepareStartService(Context context) {
         Intent intent = getStartServiceIntent(context);
+
+        // TODO: Handle this?!
+//        if (mAuthenticationType == VpnProfile.TYPE_KEYSTORE || mAuthenticationType == VpnProfile.TYPE_USERPASS_KEYSTORE) {
+//            if (getKeyStoreCertificates(context) == null)
+//                return null;
+//        }
+
         return intent;
     }
 
@@ -513,29 +760,30 @@ public class VpnProfile implements Serializable, Cloneable {
         cfg.write(getConfigFile(context, false));
         cfg.flush();
         cfg.close();
+
     }
 
     public Intent getStartServiceIntent(Context context) {
         String prefix = context.getPackageName();
+
         Intent intent = new Intent(context, OpenVPNService.class);
         intent.putExtra(prefix + ".profileUUID", mUuid.toString());
+        intent.putExtra(prefix + ".profileVersion", mVersion);
         return intent;
-    }
-
-    private String[] getKeyStoreCertificates(Context context) {
-        return getKeyStoreCertificates(context, 5);
     }
 
     public void checkForRestart(final Context context) {
         /* This method is called when OpenVPNService is restarted */
-        if ((mAuthenticationType == VpnProfile.TYPE_KEYSTORE || mAuthenticationType == VpnProfile.TYPE_USERPASS_KEYSTORE) && mPrivateKey == null) {
+
+        if ((mAuthenticationType == VpnProfile.TYPE_KEYSTORE || mAuthenticationType == VpnProfile.TYPE_USERPASS_KEYSTORE)
+                && mPrivateKey == null) {
             new Thread(new Runnable() {
                 @Override
                 public void run() {
-                    getKeyStoreCertificates(context);
+                    getExternalCertificates(context);
+
                 }
             }).start();
-
         }
     }
 
@@ -557,26 +805,59 @@ public class VpnProfile implements Serializable, Cloneable {
             VpnProfile copy = clone();
             copy.mName = name;
             return copy;
+
         } catch (CloneNotSupportedException e) {
             e.printStackTrace();
             return null;
-
         }
     }
 
-    private synchronized String[] getKeyStoreCertificates(Context context, int tries) {
+    public void pwDidFail(Context c) {
+
+    }
+
+    private X509Certificate[] getKeyStoreCertificates(Context context) throws KeyChainException, InterruptedException {
+        PrivateKey privateKey = KeyChain.getPrivateKey(context, mAlias);
+        mPrivateKey = privateKey;
+
+
+        X509Certificate[] caChain = KeyChain.getCertificateChain(context, mAlias);
+        return caChain;
+    }
+
+    private X509Certificate[] getExtAppCertificates(Context context) throws KeyChainException {
+        if (mExternalAuthenticator == null || mAlias == null)
+            throw new KeyChainException("Alias or external auth provider name not set");
+        return ExtAuthHelper.getCertificateChain(context, mExternalAuthenticator, mAlias);
+    }
+
+    public String[] getExternalCertificates(Context context) {
+        return getExternalCertificates(context, 5);
+    }
+
+
+    synchronized String[] getExternalCertificates(Context context, int tries) {
         // Force application context- KeyChain methods will block long enough that by the time they
         // are finished and try to unbind, the original activity context might have been destroyed.
         context = context.getApplicationContext();
+
         try {
-            mPrivateKey = KeyChain.getPrivateKey(context, mAlias);
             String keystoreChain = null;
-            X509Certificate[] caChain = KeyChain.getCertificateChain(context, mAlias);
-            if (caChain == null) throw new NoCertReturnedException("No certificate returned from Keystore");
+
+            X509Certificate caChain[];
+            if (mAuthenticationType == TYPE_EXTERNAL_APP) {
+                caChain = getExtAppCertificates(context);
+            } else {
+                caChain = getKeyStoreCertificates(context);
+            }
+            if (caChain == null)
+                throw new NoCertReturnedException("No certificate returned from Keystore");
+
             if (caChain.length <= 1 && TextUtils.isEmpty(mCaFilename)) {
                 VpnStatus.logMessage(VpnStatus.LogLevel.ERROR, "", context.getString(R.string.keychain_nocacert));
             } else {
                 StringWriter ksStringWriter = new StringWriter();
+
                 PemWriter pw = new PemWriter(ksStringWriter);
                 for (int i = 1; i < caChain.length; i++) {
                     X509Certificate cert = caChain[i];
@@ -584,30 +865,41 @@ public class VpnProfile implements Serializable, Cloneable {
                 }
                 pw.close();
                 keystoreChain = ksStringWriter.toString();
-
             }
+
+
             String caout = null;
             if (!TextUtils.isEmpty(mCaFilename)) {
                 try {
                     Certificate[] cacerts = X509Utils.getCertificatesFromFile(mCaFilename);
                     StringWriter caoutWriter = new StringWriter();
                     PemWriter pw = new PemWriter(caoutWriter);
+
                     for (Certificate cert : cacerts)
                         pw.writeObject(new PemObject("CERTIFICATE", cert.getEncoded()));
                     pw.close();
                     caout = caoutWriter.toString();
+
                 } catch (Exception e) {
                     VpnStatus.logError("Could not read CA certificate" + e.getLocalizedMessage());
                 }
             }
+
+
             StringWriter certout = new StringWriter();
+
+
             if (caChain.length >= 1) {
                 X509Certificate usercert = caChain[0];
+
                 PemWriter upw = new PemWriter(certout);
                 upw.writeObject(new PemObject("CERTIFICATE", usercert.getEncoded()));
                 upw.close();
+
             }
             String user = certout.toString();
+
+
             String ca, extra;
             if (caout == null) {
                 ca = keystoreChain;
@@ -616,11 +908,13 @@ public class VpnProfile implements Serializable, Cloneable {
                 ca = caout;
                 extra = keystoreChain;
             }
+
             return new String[]{ca, extra, user};
         } catch (InterruptedException | IOException | KeyChainException | NoCertReturnedException | IllegalArgumentException
                 | CertificateException e) {
             e.printStackTrace();
             VpnStatus.logError(R.string.keyChainAccessError, e.getLocalizedMessage());
+
             VpnStatus.logError(R.string.keychain_access);
             if (Build.VERSION.SDK_INT == Build.VERSION_CODES.JELLY_BEAN) {
                 if (!mAlias.matches("^[a-zA-Z0-9]$")) {
@@ -628,57 +922,115 @@ public class VpnProfile implements Serializable, Cloneable {
                 }
             }
             return null;
+
         } catch (AssertionError e) {
-            if (tries == 0) return null;
+            if (tries == 0)
+                return null;
             VpnStatus.logError(String.format("Failure getting Keystore Keys (%s), retrying", e.getLocalizedMessage()));
             try {
                 Thread.sleep(3000);
             } catch (InterruptedException e1) {
                 VpnStatus.logException(e1);
             }
-            return getKeyStoreCertificates(context, tries - 1);
+            return getExternalCertificates(context, tries - 1);
         }
+
+    }
+
+    public int checkProfile(Context c) {
+        return checkProfile(c, doUseOpenVPN3(c));
     }
 
     //! Return an error if something is wrong
-    int checkProfile(Context context) {
-        if (mAuthenticationType == TYPE_KEYSTORE || mAuthenticationType == TYPE_USERPASS_KEYSTORE) {
-            if (mAlias == null) return R.string.no_keystore_cert_selected;
+    public int checkProfile(Context context, boolean useOpenVPN3) {
+        if (mAuthenticationType == TYPE_KEYSTORE || mAuthenticationType == TYPE_USERPASS_KEYSTORE || mAuthenticationType == TYPE_EXTERNAL_APP) {
+            if (mAlias == null)
+                return R.string.no_keystore_cert_selected;
+        } else if (mAuthenticationType == TYPE_CERTIFICATES || mAuthenticationType == TYPE_USERPASS_CERTIFICATES) {
+            if (TextUtils.isEmpty(mCaFilename))
+                return R.string.no_ca_cert_selected;
         }
+
+        if (mCheckRemoteCN && mX509AuthType == X509_VERIFY_TLSREMOTE)
+            return R.string.deprecated_tls_remote;
+
         if (!mUsePull || mAuthenticationType == TYPE_STATICKEYS) {
-            if (mIPv4Address == null || cidrToIPAndNetmask(mIPv4Address) == null) return R.string.ipv4_format_error;
+            if (mIPv4Address == null || cidrToIPAndNetmask(mIPv4Address) == null)
+                return R.string.ipv4_format_error;
         }
         if (!mUseDefaultRoute) {
-            if (!TextUtils.isEmpty(mCustomRoutes) && getCustomRoutes(mCustomRoutes).size() == 0) return R.string.custom_route_format_error;
-            if (!TextUtils.isEmpty(mExcludedRoutes) && getCustomRoutes(mExcludedRoutes).size() == 0) return R.string.custom_route_format_error;
+            if (!TextUtils.isEmpty(mCustomRoutes) && getCustomRoutes(mCustomRoutes).size() == 0)
+                return R.string.custom_route_format_error;
+
+            if (!TextUtils.isEmpty(mExcludedRoutes) && getCustomRoutes(mExcludedRoutes).size() == 0)
+                return R.string.custom_route_format_error;
+
         }
-        if (mUseTLSAuth && TextUtils.isEmpty(mTLSAuthFilename)) return R.string.missing_tlsauth;
-        if ((mAuthenticationType == TYPE_USERPASS_CERTIFICATES || mAuthenticationType == TYPE_CERTIFICATES) && (TextUtils.isEmpty(mClientCertFilename) || TextUtils.isEmpty(mClientKeyFilename)))
+
+        if (mUseTLSAuth && TextUtils.isEmpty(mTLSAuthFilename))
+            return R.string.missing_tlsauth;
+
+        if ((mAuthenticationType == TYPE_USERPASS_CERTIFICATES || mAuthenticationType == TYPE_CERTIFICATES)
+                && (TextUtils.isEmpty(mClientCertFilename) || TextUtils.isEmpty(mClientKeyFilename)))
             return R.string.missing_certificates;
-        if ((mAuthenticationType == TYPE_CERTIFICATES || mAuthenticationType == TYPE_USERPASS_CERTIFICATES) && TextUtils.isEmpty(mCaFilename)) return R.string.missing_ca_certificate;
+
+        if ((mAuthenticationType == TYPE_CERTIFICATES || mAuthenticationType == TYPE_USERPASS_CERTIFICATES)
+                && TextUtils.isEmpty(mCaFilename))
+            return R.string.missing_ca_certificate;
+
+
         boolean noRemoteEnabled = true;
-        for (Connection c : mConnections)
-            if (c.mEnabled) noRemoteEnabled = false;
-        if (noRemoteEnabled) return R.string.remote_no_server_selected;
+        for (Connection c : mConnections) {
+            if (c.mEnabled)
+                noRemoteEnabled = false;
+
+        }
+        if (noRemoteEnabled)
+            return R.string.remote_no_server_selected;
+
+        if (useOpenVPN3) {
+            if (mAuthenticationType == TYPE_STATICKEYS) {
+                return R.string.openvpn3_nostatickeys;
+            }
+            if (mAuthenticationType == TYPE_PKCS12 || mAuthenticationType == TYPE_USERPASS_PKCS12) {
+                return R.string.openvpn3_pkcs12;
+            }
+            for (Connection conn : mConnections) {
+                if (conn.mProxyType == Connection.ProxyType.ORBOT || conn.mProxyType == Connection.ProxyType.SOCKS5)
+                    return R.string.openvpn3_socksproxy;
+            }
+        }
+        for (Connection c : mConnections) {
+            if (c.mProxyType == Connection.ProxyType.ORBOT) {
+                if (usesExtraProxyOptions())
+                    return R.string.error_orbot_and_proxy_options;
+                if (!OrbotHelper.checkTorReceier(context))
+                    return R.string.no_orbotfound;
+            }
+        }
+
+
         // Everything okay
         return R.string.no_error_found;
+
     }
 
     //! Openvpn asks for a "Private Key", this should be pkcs12 key
     //
     public String getPasswordPrivateKey() {
-        if (mTransientPCKS12PW != null) {
-            String pwcopy = mTransientPCKS12PW;
-            mTransientPCKS12PW = null;
-            return pwcopy;
+        String cachedPw = PasswordCache.getPKCS12orCertificatePassword(mUuid, true);
+        if (cachedPw != null) {
+            return cachedPw;
         }
         switch (mAuthenticationType) {
             case TYPE_PKCS12:
             case TYPE_USERPASS_PKCS12:
                 return mPKCS12Password;
+
             case TYPE_CERTIFICATES:
             case TYPE_USERPASS_CERTIFICATES:
                 return mKeyPassword;
+
             case TYPE_USERPASS:
             case TYPE_STATICKEYS:
             default:
@@ -686,7 +1038,7 @@ public class VpnProfile implements Serializable, Cloneable {
         }
     }
 
-    private boolean isUserPWAuth() {
+    public boolean isUserPWAuth() {
         switch (mAuthenticationType) {
             case TYPE_USERPASS:
             case TYPE_USERPASS_CERTIFICATES:
@@ -695,13 +1047,17 @@ public class VpnProfile implements Serializable, Cloneable {
                 return true;
             default:
                 return false;
+
         }
     }
 
-    private boolean requireTLSKeyPassword() {
-        if (TextUtils.isEmpty(mClientKeyFilename)) return false;
+    public boolean requireTLSKeyPassword() {
+        if (TextUtils.isEmpty(mClientKeyFilename))
+            return false;
+
         String data = "";
-        if (isEmbedded(mClientKeyFilename)) data = mClientKeyFilename;
+        if (isEmbedded(mClientKeyFilename))
+            data = mClientKeyFilename;
         else {
             char[] buf = new char[2048];
             FileReader fr;
@@ -718,30 +1074,43 @@ public class VpnProfile implements Serializable, Cloneable {
             } catch (IOException e) {
                 return false;
             }
+
         }
-        return data.contains("Proc-Type: 4,ENCRYPTED") || data.contains("-----BEGIN ENCRYPTED PRIVATE KEY-----");
+
+        if (data.contains("Proc-Type: 4,ENCRYPTED"))
+            return true;
+        else if (data.contains("-----BEGIN ENCRYPTED PRIVATE KEY-----"))
+            return true;
+        else
+            return false;
     }
 
-    public int needUserPWInput(boolean ignoreTransient) {
-        if ((mAuthenticationType == TYPE_PKCS12 || mAuthenticationType == TYPE_USERPASS_PKCS12) && (mPKCS12Password == null || mPKCS12Password.equals(""))) {
-            if (ignoreTransient || mTransientPCKS12PW == null) return R.string.pkcs12_file_encryption_key;
+    public int needUserPWInput(String transientCertOrPkcs12PW, String mTransientAuthPW) {
+        if ((mAuthenticationType == TYPE_PKCS12 || mAuthenticationType == TYPE_USERPASS_PKCS12) &&
+                (mPKCS12Password == null || mPKCS12Password.equals(""))) {
+            if (transientCertOrPkcs12PW == null)
+                return R.string.pkcs12_file_encryption_key;
         }
+
         if (mAuthenticationType == TYPE_CERTIFICATES || mAuthenticationType == TYPE_USERPASS_CERTIFICATES) {
-            if (requireTLSKeyPassword() && TextUtils.isEmpty(mKeyPassword)) if (ignoreTransient || mTransientPCKS12PW == null) {
-                return R.string.private_key_password;
-            }
+            if (requireTLSKeyPassword() && TextUtils.isEmpty(mKeyPassword))
+                if (transientCertOrPkcs12PW == null) {
+                    return R.string.private_key_password;
+                }
         }
-        if (isUserPWAuth() && (TextUtils.isEmpty(mUsername) || (TextUtils.isEmpty(mPassword) && (mTransientPW == null || ignoreTransient)))) {
+
+        if (isUserPWAuth() &&
+                (TextUtils.isEmpty(mUsername) ||
+                        (TextUtils.isEmpty(mPassword) && mTransientAuthPW == null))) {
             return R.string.password;
         }
         return 0;
     }
 
     public String getPasswordAuth() {
-        if (mTransientPW != null) {
-            String pwcopy = mTransientPW;
-            mTransientPW = null;
-            return pwcopy;
+        String cachedPw = PasswordCache.getAuthPassword(mUuid, true);
+        if (cachedPw != null) {
+            return cachedPw;
         } else {
             return mPassword;
         }
@@ -754,60 +1123,129 @@ public class VpnProfile implements Serializable, Cloneable {
     }
 
     public String getUUIDString() {
-        return mUuid.toString();
+        return mUuid.toString().toLowerCase(Locale.ENGLISH);
     }
 
-    private PrivateKey getKeystoreKey() {
+    public PrivateKey getKeystoreKey() {
         return mPrivateKey;
     }
 
-    public String getSignedData(String b64data) {
-        PrivateKey privkey = getKeystoreKey();
+    @Nullable
+    public String getSignedData(Context c, String b64data, boolean pkcs1padding) {
         byte[] data = Base64.decode(b64data, Base64.DEFAULT);
+        byte[] signed_bytes;
+        if (mAuthenticationType == TYPE_EXTERNAL_APP)
+            signed_bytes = getExtAppSignedData(c, data);
+        else
+            signed_bytes = getKeyChainSignedData(data, pkcs1padding);
+
+        if (signed_bytes != null)
+            return Base64.encodeToString(signed_bytes, Base64.NO_WRAP);
+        else
+            return null;
+    }
+
+    private byte[] getExtAppSignedData(Context c, byte[] data) {
+        if (TextUtils.isEmpty(mExternalAuthenticator))
+            return null;
+        try {
+            return ExtAuthHelper.signData(c, mExternalAuthenticator, mAlias, data);
+        } catch (KeyChainException | InterruptedException e) {
+            VpnStatus.logError(R.string.error_extapp_sign, mExternalAuthenticator, e.getClass().toString(), e.getLocalizedMessage());
+            return null;
+        }
+    }
+
+    private byte[] getKeyChainSignedData(byte[] data, boolean pkcs1padding) {
+
+        PrivateKey privkey = getKeystoreKey();
         // The Jelly Bean *evil* Hack
         // 4.2 implements the RSA/ECB/PKCS1PADDING in the OpenSSLprovider
         if (Build.VERSION.SDK_INT == Build.VERSION_CODES.JELLY_BEAN) {
-            return processSignJellyBeans(privkey, data);
+            return processSignJellyBeans(privkey, data, pkcs1padding);
         }
+
+
         try {
+            @SuppressLint("GetInstance")
+            String keyalgorithm = privkey.getAlgorithm();
+
+            byte[] signed_bytes;
+            if (keyalgorithm.equals("EC")) {
+                Signature signer = Signature.getInstance("NONEwithECDSA");
+
+                signer.initSign(privkey);
+                signer.update(data);
+                signed_bytes = signer.sign();
+
+            } else {
             /* ECB is perfectly fine in this special case, since we are using it for
                the public/private part in the TLS exchange
              */
-            @SuppressLint("GetInstance") Cipher rsaSigner = Cipher.getInstance("RSA/ECB/PKCS1PADDING");
-            rsaSigner.init(Cipher.ENCRYPT_MODE, privkey);
-            byte[] signed_bytes = rsaSigner.doFinal(data);
-            return Base64.encodeToString(signed_bytes, Base64.NO_WRAP);
+                Cipher signer;
+                if (pkcs1padding)
+                    signer = Cipher.getInstance("RSA/ECB/PKCS1PADDING");
+                else
+                    signer = Cipher.getInstance("RSA/ECB/NoPadding");
+
+
+                signer.init(Cipher.ENCRYPT_MODE, privkey);
+
+                signed_bytes = signer.doFinal(data);
+            }
+            return signed_bytes;
         } catch (NoSuchAlgorithmException | InvalidKeyException | IllegalBlockSizeException
-                | BadPaddingException | NoSuchPaddingException e) {
+                | BadPaddingException | NoSuchPaddingException | SignatureException e) {
             VpnStatus.logError(R.string.error_rsa_sign, e.getClass().toString(), e.getLocalizedMessage());
             return null;
         }
     }
 
-    private String processSignJellyBeans(PrivateKey privkey, byte[] data) {
+    private byte[] processSignJellyBeans(PrivateKey privkey, byte[] data, boolean pkcs1padding) {
         try {
             Method getKey = privkey.getClass().getSuperclass().getDeclaredMethod("getOpenSSLKey");
             getKey.setAccessible(true);
+
             // Real object type is OpenSSLKey
             Object opensslkey = getKey.invoke(privkey);
+
             getKey.setAccessible(false);
+
             Method getPkeyContext = opensslkey.getClass().getDeclaredMethod("getPkeyContext");
+
             // integer pointer to EVP_pkey
             getPkeyContext.setAccessible(true);
             int pkey = (Integer) getPkeyContext.invoke(opensslkey);
             getPkeyContext.setAccessible(false);
+
             // 112 with TLS 1.2 (172 back with 4.3), 36 with TLS 1.0
-            byte[] signed_bytes = NativeUtils.rsasign(data, pkey);
-            return Base64.encodeToString(signed_bytes, Base64.NO_WRAP);
+            return NativeUtils.rsasign(data, pkey, pkcs1padding);
+
         } catch (NoSuchMethodException | InvalidKeyException | InvocationTargetException | IllegalAccessException | IllegalArgumentException e) {
             VpnStatus.logError(R.string.error_rsa_sign, e.getClass().toString(), e.getLocalizedMessage());
             return null;
         }
     }
 
-    private class NoCertReturnedException extends Exception {
-        NoCertReturnedException(String msg) {
+    private boolean usesExtraProxyOptions() {
+        if (mUseCustomConfig && mCustomConfigOptions != null && mCustomConfigOptions.contains("http-proxy-option "))
+            return true;
+        for (Connection c : mConnections)
+            if (c.usesExtraProxyOptions())
+                return true;
+
+        return false;
+    }
+
+    class NoCertReturnedException extends Exception {
+        public NoCertReturnedException(String msg) {
             super(msg);
         }
     }
+
+
 }
+
+
+
+
