@@ -2,37 +2,92 @@
  * Copyright (c) 2012-2016 Arne Schwabe
  * Distributed under the GNU GPL v2 with additional terms. For full terms see the file doc/LICENSE.txt
  */
+
 package de.blinkt.openvpn.core;
 
+import android.annotation.TargetApi;
 import android.app.Application;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.content.Context;
+import android.graphics.Color;
+import android.os.Build;
 
+import android.os.StrictMode;
 import de.blinkt.openvpn.BuildConfig;
+import de.blinkt.openvpn.R;
+import de.blinkt.openvpn.api.AppRestrictions;
 
-/*
-import org.acra.ACRA;
-import org.acra.ReportingInteractionMode;
-import org.acra.annotation.ReportsCrashes;
-*/
-/*
-@ReportsCrashes(
-        formKey = "",
-        formUri = "http://reports.blinkt.de/report-icsopenvpn",
-        reportType = org.acra.sender.HttpSender.Type.JSON,
-        httpMethod = org.acra.sender.HttpSender.Method.PUT,
-        formUriBasicAuthLogin="report-icsopenvpn",
-        formUriBasicAuthPassword="Tohd4neiF9Ai!!!!111eleven",
-        mode = ReportingInteractionMode.TOAST,
-        resToastText = R.string.crash_toast_text
-)
-*/
 public class ICSOpenVPNApplication extends Application {
+    private StatusListener mStatus;
+
     @Override
     public void onCreate() {
+        if("robolectric".equals(Build.FINGERPRINT))
+            return;
+
         super.onCreate();
         PRNGFixes.apply();
-        if (BuildConfig.DEBUG) {
-            //ACRA.init(this);
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
+            createNotificationChannels();
+        mStatus = new StatusListener();
+        mStatus.init(getApplicationContext());
+
+        if (BuildConfig.BUILD_TYPE.equals("debug"))
+            enableStrictModes();
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            AppRestrictions.getInstance(this).checkRestrictions(this);
         }
-        VpnStatus.initLogCache(getApplicationContext().getCacheDir());
+    }
+
+    private void enableStrictModes() {
+        StrictMode.VmPolicy policy = new StrictMode.VmPolicy.Builder()
+                .detectAll()
+                .penaltyLog()
+                .penaltyDeath()
+                .build();
+        StrictMode.setVmPolicy(policy);
+
+    }
+
+    @TargetApi(Build.VERSION_CODES.O)
+    private void createNotificationChannels() {
+        NotificationManager mNotificationManager =
+                (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+
+        // Background message
+        CharSequence name = getString(R.string.channel_name_background);
+        NotificationChannel mChannel = new NotificationChannel(OpenVPNService.NOTIFICATION_CHANNEL_BG_ID,
+                name, NotificationManager.IMPORTANCE_MIN);
+
+        mChannel.setDescription(getString(R.string.channel_description_background));
+        mChannel.enableLights(false);
+
+        mChannel.setLightColor(Color.DKGRAY);
+        mNotificationManager.createNotificationChannel(mChannel);
+
+        // Connection status change messages
+
+        name = getString(R.string.channel_name_status);
+        mChannel = new NotificationChannel(OpenVPNService.NOTIFICATION_CHANNEL_NEWSTATUS_ID,
+                name, NotificationManager.IMPORTANCE_LOW);
+
+        mChannel.setDescription(getString(R.string.channel_description_status));
+        mChannel.enableLights(true);
+
+        mChannel.setLightColor(Color.BLUE);
+        mNotificationManager.createNotificationChannel(mChannel);
+
+
+        // Urgent requests, e.g. two factor auth
+        name = getString(R.string.channel_name_userreq);
+        mChannel = new NotificationChannel(OpenVPNService.NOTIFICATION_CHANNEL_USERREQ_ID,
+                name, NotificationManager.IMPORTANCE_HIGH);
+        mChannel.setDescription(getString(R.string.channel_description_userreq));
+        mChannel.enableVibration(true);
+        mChannel.setLightColor(Color.CYAN);
+        mNotificationManager.createNotificationChannel(mChannel);
     }
 }
