@@ -2,9 +2,20 @@
  * Copyright (c) 2012-2016 Arne Schwabe
  * Distributed under the GNU GPL v2 with additional terms. For full terms see the file doc/LICENSE.txt
  */
+
 package de.blinkt.openvpn.core;
 
 public interface OpenVPNManagement {
+    interface PausedStateCallback {
+        boolean shouldBeRunning();
+    }
+
+    enum pauseReason {
+        noNetwork,
+        userPause,
+        screenOff,
+    }
+
     int mBytecountInterval = 2;
 
     void reconnect();
@@ -17,8 +28,7 @@ public interface OpenVPNManagement {
      * @param replaceConnection True if the VPN is connected by a new connection.
      * @return true if there was a process that has been send a stop signal
      */
-    boolean
-    stopVPN(boolean replaceConnection);
+    boolean stopVPN(boolean replaceConnection);
 
     /*
      * Rebind the interface
@@ -27,13 +37,9 @@ public interface OpenVPNManagement {
 
     void setPauseCallback(PausedStateCallback callback);
 
-    enum pauseReason {
-        noNetwork,
-        userPause,
-        screenOff,
-    }
-
-    interface PausedStateCallback {
-        boolean shouldBeRunning();
-    }
+    /**
+     * Send the response to a challenge response
+     * @param response  Base64 encoded response
+     */
+    void sendCRResponse(String response);
 }
