@@ -2,6 +2,7 @@
  * Copyright (c) 2012-2016 Arne Schwabe
  * Distributed under the GNU GPL v2 with additional terms. For full terms see the file doc/LICENSE.txt
  */
+
 package de.blinkt.openvpn.core;
 
 import android.Manifest.permission;
@@ -16,22 +17,25 @@ import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
+import android.content.pm.ShortcutManager;
 import android.content.res.Configuration;
+import android.content.res.Resources;
 import android.graphics.Color;
 import android.net.ConnectivityManager;
 import android.net.VpnService;
 import android.os.Binder;
 import android.os.Build;
+import android.os.Bundle;
 import android.os.Handler;
 import android.os.Handler.Callback;
 import android.os.IBinder;
 import android.os.Message;
 import android.os.ParcelFileDescriptor;
-import android.preference.PreferenceManager;
+import android.os.RemoteException;
 import android.system.OsConstants;
 import android.text.TextUtils;
+import android.util.Base64;
 import android.util.Log;
 import android.widget.Toast;
 
@@ -45,53 +49,57 @@ import java.lang.reflect.Method;
 import java.net.Inet6Address;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
-import java.text.SimpleDateFormat;
+import java.nio.charset.Charset;
 import java.util.Calendar;
 import java.util.Collection;
-import java.util.Date;
 import java.util.Locale;
+import java.util.Objects;
 import java.util.Vector;
 
-import de.blinkt.openvpn.BuildConfig;
 import de.blinkt.openvpn.DisconnectVPNActivity;
+import de.blinkt.openvpn.LaunchVPN;
 import de.blinkt.openvpn.R;
 import de.blinkt.openvpn.VpnProfile;
+import de.blinkt.openvpn.api.ExternalAppDatabase;
 import de.blinkt.openvpn.core.VpnStatus.ByteCountListener;
-import de.blinkt.openvpn.core.VpnStatus.ConnectionStatus;
 import de.blinkt.openvpn.core.VpnStatus.StateListener;
+import de.blinkt.openvpn.utils.TotalTraffic;
 
-import static android.app.Notification.PRIORITY_DEFAULT;
-import static android.app.Notification.PRIORITY_MAX;
-import static android.app.Notification.PRIORITY_MIN;
-import static de.blinkt.openvpn.core.NetworkSpace.ipAddress;
-import static de.blinkt.openvpn.core.VpnStatus.ConnectionStatus.LEVEL_CONNECTED;
-import static de.blinkt.openvpn.core.VpnStatus.ConnectionStatus.LEVEL_WAITING_FOR_USER_INPUT;
+import static de.blinkt.openvpn.core.ConnectionStatus.LEVEL_CONNECTED;
+import static de.blinkt.openvpn.core.ConnectionStatus.LEVEL_WAITING_FOR_USER_INPUT;
+import static de.blinkt.openvpn.core.NetworkSpace.IpAddress;
 
-public class OpenVPNService extends VpnService implements StateListener, Callback, ByteCountListener {
+public class OpenVPNService extends VpnService implements StateListener, Callback, ByteCountListener, IOpenVPNServiceInternal {
 
     private String byteIn, byteOut;
     private String duration;
-    private String lastChannel;
 
-    public static final String NOTIFICATION_CHANNEL_BG_ID = "openvpn_bg";
-    public static final String NOTIFICATION_CHANNEL_NEWSTATUS_ID = "openvpn_notification";
-    public static final String NOTIFICATION_CHANNEL_USERREQ_ID = "openvpn_userreq";
-
-    public static final String TAG = "VpnService";
     public static final String START_SERVICE = "de.blinkt.openvpn.START_SERVICE";
     public static final String START_SERVICE_STICKY = "de.blinkt.openvpn.START_SERVICE_STICKY";
     public static final String ALWAYS_SHOW_NOTIFICATION = "de.blinkt.openvpn.NOTIFICATION_ALWAYS_VISIBLE";
     public static final String DISCONNECT_VPN = "de.blinkt.openvpn.DISCONNECT_VPN";
+    public static final String NOTIFICATION_CHANNEL_BG_ID = "openvpn_bg";
+    public static final String NOTIFICATION_CHANNEL_NEWSTATUS_ID = "openvpn_newstat";
+    public static final String NOTIFICATION_CHANNEL_USERREQ_ID = "openvpn_userreq";
+
+    public static final String VPNSERVICE_TUN = "vpnservice-tun";
+    public final static String ORBOT_PACKAGE_NAME = "org.torproject.android";
     private static final String PAUSE_VPN = "de.blinkt.openvpn.PAUSE_VPN";
     private static final String RESUME_VPN = "de.blinkt.openvpn.RESUME_VPN";
-    private static final int OPENVPN_STATUS = 1;
+
+    public static final String EXTRA_CHALLENGE_TXT = "de.blinkt.openvpn.core.CR_TEXT_CHALLENGE";
+    public static final String EXTRA_CHALLENGE_OPENURL = "de.blinkt.openvpn.core.OPENURL_CHALLENGE";
+
+    private static final int PRIORITY_MIN = -2;
+    private static final int PRIORITY_DEFAULT = 0;
+    private static final int PRIORITY_MAX = 2;
     private static boolean mNotificationAlwaysVisible = false;
-    private static Class mNotificationActivityClass;
+    private static Class<? extends Activity> mNotificationActivityClass;
     private final Vector<String> mDnslist = new Vector<>();
     private final NetworkSpace mRoutes = new NetworkSpace();
     private final NetworkSpace mRoutesv6 = new NetworkSpace();
-    private final IBinder mBinder = new LocalBinder();
     private final Object mProcessLock = new Object();
+    private String lastChannel;
     private Thread mProcessThread = null;
     private VpnProfile mProfile;
     private String mDomain = null;
@@ -102,32 +110,86 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
     private boolean mDisplayBytecount = false;
     private boolean mStarting = false;
     private long mConnecttime;
-    private boolean mOvpn3 = false;
     private OpenVPNManagement mManagement;
+    /*private final IBinder mBinder = new IOpenVPNServiceInternal.Stub() {
+
+        @Override
+        public boolean protect(int fd) throws RemoteException {
+            return OpenVPNService.this.protect(fd);
+        }
+
+        @Override
+        public void userPause(boolean shouldbePaused) throws RemoteException {
+            OpenVPNService.this.userPause(shouldbePaused);
+        }
+
+        @Override
+        public boolean stopVPN(boolean replaceConnection) throws RemoteException {
+            return OpenVPNService.this.stopVPN(replaceConnection);
+        }
+
+        @Override
+        public void addAllowedExternalApp(String packagename) throws RemoteException {
+            OpenVPNService.this.addAllowedExternalApp(packagename);
+        }
+
+        @Override
+        public boolean isAllowedExternalApp(String packagename) throws RemoteException {
+            return OpenVPNService.this.isAllowedExternalApp(packagename);
+
+        }
+
+        @Override
+        public void challengeResponse(String repsonse) throws RemoteException {
+            OpenVPNService.this.challengeResponse(repsonse);
+        }
+
+
+    };*/
+
+    private final IBinder mBinder = new LocalBinder();
+    private static String state = "";
+    boolean flag = false;
     private String mLastTunCfg;
     private String mRemoteGW;
     private Handler guiHandler;
     private Toast mlastToast;
     private Runnable mOpenVPNThread;
-    boolean flag = false;
 
-    private static String state = "";
-
-    public static String humanReadableByteCount(long bytes, boolean mbit) {
-        if (mbit) {
+    // From: http://stackoverflow.com/questions/3758606/how-to-convert-byte-size-into-human-readable-format-in-java
+    public static String humanReadableByteCount(long bytes, boolean speed, Resources res) {
+        if (speed)
             bytes = bytes * 8;
-        }
-        int unit = mbit ? 1000 : 1024;
-        if (bytes < unit) {
-            return bytes + (mbit ? " bit" : " B");
-        }
-        int exp = (int) (Math.log(bytes) / Math.log(unit));
-        String pre = (mbit ? "kMGTPE" : "KMGTPE").charAt(exp - 1) + (mbit ? "" : "");
-        if (mbit) {
-            return String.format(Locale.getDefault(), "%.1f %sbit", bytes / Math.pow(unit, exp), pre);
-        } else {
-            return String.format(Locale.getDefault(), "%.1f %sB", bytes / Math.pow(unit, exp), pre);
-        }
+        int unit = speed ? 1000 : 1024;
+
+
+        int exp = Math.max(0, Math.min((int) (Math.log(bytes) / Math.log(unit)), 3));
+
+        float bytesUnit = (float) (bytes / Math.pow(unit, exp));
+
+        if (speed)
+            switch (exp) {
+                case 0:
+                    return res.getString(R.string.bits_per_second, bytesUnit);
+                case 1:
+                    return res.getString(R.string.kbits_per_second, bytesUnit);
+                case 2:
+                    return res.getString(R.string.mbits_per_second, bytesUnit);
+                default:
+                    return res.getString(R.string.gbits_per_second, bytesUnit);
+            }
+        else
+            switch (exp) {
+                case 0:
+                    return res.getString(R.string.volume_byte, bytesUnit);
+                case 1:
+                    return res.getString(R.string.volume_kbyte, bytesUnit);
+                case 2:
+                    return res.getString(R.string.volume_mbyte, bytesUnit);
+                default:
+                    return res.getString(R.string.volume_gbyte, bytesUnit);
+
+            }
     }
 
     /**
@@ -139,11 +201,54 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         mNotificationActivityClass = activityClass;
     }
 
+    PendingIntent getContentIntent() {
+        try {
+            if (mNotificationActivityClass != null) {
+                // Let the configure Button show the Log
+                Intent intent = new Intent(getBaseContext(), mNotificationActivityClass);
+                String typeStart = Objects.requireNonNull(
+                        mNotificationActivityClass.getField("TYPE_START").get(null)).toString();
+                Integer typeFromNotify = Integer.parseInt(Objects.requireNonNull(mNotificationActivityClass.getField("TYPE_FROM_NOTIFY").get(null)).toString());
+                intent.putExtra(typeStart, typeFromNotify);
+                intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK |
+                        Intent.FLAG_ACTIVITY_SINGLE_TOP);
+                return PendingIntent.getActivity(this, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
+            }
+        } catch (Exception e) {
+            Log.e(this.getClass().getCanonicalName(), "Build detail intent error", e);
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    @Override
+    public void addAllowedExternalApp(String packagename) throws RemoteException {
+        ExternalAppDatabase extapps = new ExternalAppDatabase(OpenVPNService.this);
+        extapps.addApp(packagename);
+    }
+
+    @Override
+    public boolean isAllowedExternalApp(String packagename) throws RemoteException {
+        ExternalAppDatabase extapps = new ExternalAppDatabase(OpenVPNService.this);
+        return extapps.checkRemoteActionPermission(this, packagename);
+    }
+
+    @Override
+    public void challengeResponse(String response) throws RemoteException {
+        if (mManagement != null) {
+            String b64response = Base64.encodeToString(response.getBytes(Charset.forName("UTF-8")), Base64.DEFAULT);
+            mManagement.sendCRResponse(b64response);
+        }
+    }
+
+
     @Override
     public IBinder onBind(Intent intent) {
         String action = intent.getAction();
-        if (action != null && action.equals(START_SERVICE)) return mBinder;
-        else return super.onBind(intent);
+        if (action != null && action.equals(START_SERVICE))
+            return mBinder;
+        else
+            return super.onBind(intent);
     }
 
     @Override
@@ -154,7 +259,7 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
     }
 
     // Similar to revoke but do not try to stop process
-    public void processDied() {
+    public void openvpnStopped() {
         endVpnService();
     }
 
@@ -168,6 +273,7 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         mOpenVPNThread = null;
         if (!mStarting) {
             stopForeground(!mNotificationAlwaysVisible);
+
             if (!mNotificationAlwaysVisible) {
                 stopSelf();
                 VpnStatus.removeStateListener(this);
@@ -175,10 +281,28 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         }
     }
 
+    @RequiresApi(Build.VERSION_CODES.O)
+    private String createNotificationChannel(String channelId, String channelName) {
+        NotificationChannel chan = new NotificationChannel(channelId,
+                channelName, NotificationManager.IMPORTANCE_NONE);
+        chan.setLightColor(Color.BLUE);
+        chan.setLockscreenVisibility(Notification.VISIBILITY_PRIVATE);
+        NotificationManager service = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+        service.createNotificationChannel(chan);
+        return channelId;
+    }
+
     private void showNotification(final String msg, String tickerText, @NonNull String channel,
                                   long when, ConnectionStatus status, Intent intent) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            channel = createNotificationChannel(channel, channel + " Name");
+        } else {
+            // If earlier version channel ID is not used
+            // https://developer.android.com/reference/android/support/v4/app/NotificationCompat.Builder.html#NotificationCompat.Builder(android.content.Context)
+            channel = "";
+        }
+
         NotificationManager mNotificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
-        int icon = getIconByConnectionStatus(status);
 
         android.app.Notification.Builder nbuilder = new Notification.Builder(this);
 
@@ -198,14 +322,17 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         nbuilder.setContentText(msg);
         nbuilder.setOnlyAlertOnce(true);
         nbuilder.setOngoing(true);
-
-
-        nbuilder.setSmallIcon(icon);
+        nbuilder.setSmallIcon(R.drawable.ic_notification);
         if (status == LEVEL_WAITING_FOR_USER_INPUT) {
             PendingIntent pIntent = PendingIntent.getActivity(this, 0, intent, 0);
             nbuilder.setContentIntent(pIntent);
         } else {
-            nbuilder.setContentIntent(getGraphPendingIntent());
+            PendingIntent contentPendingIntent = getContentIntent();
+            if (contentPendingIntent != null) {
+                nbuilder.setContentIntent(contentPendingIntent);
+            } else {
+                nbuilder.setContentIntent(getGraphPendingIntent());
+            }
         }
 
         if (when != 0)
@@ -221,40 +348,32 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
             lpNotificationExtras(nbuilder, Notification.CATEGORY_SERVICE);
 
-        // For android orio or upper version we need to create notification channel
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-
-            int importance = NotificationManager.IMPORTANCE_DEFAULT;
-            NotificationChannel notificationChannel = new NotificationChannel(channel, "VPN Service", importance);
-            // Register the channel with the system; you can't change the importance
-            // or other notification behaviors after this
-            NotificationManager notificationManager = getSystemService(NotificationManager.class);
-            notificationManager.createNotificationChannel(notificationChannel);
-
             //noinspection NewApi
             nbuilder.setChannelId(channel);
-            if (mProfile != null) {
+            if (mProfile != null)
                 //noinspection NewApi
                 nbuilder.setShortcutId(mProfile.getUUIDString());
-            }
 
         }
 
         if (tickerText != null && !tickerText.equals(""))
             nbuilder.setTicker(tickerText);
+        try {
+            Notification notification = nbuilder.build();
 
-        @SuppressWarnings("deprecation")
-        Notification notification = nbuilder.getNotification();
-        int notificationId = channel.hashCode();
+            int notificationId = channel.hashCode();
 
+            mNotificationManager.notify(notificationId, notification);
 
-        mNotificationManager.notify(notificationId, notification);
+            startForeground(notificationId, notification);
 
-        startForeground(notificationId, notification);
-
-        if (lastChannel != null && !channel.equals(lastChannel)) {
-            // Cancel old notification
-            mNotificationManager.cancel(lastChannel.hashCode());
+            if (lastChannel != null && !channel.equals(lastChannel)) {
+                // Cancel old notification
+                mNotificationManager.cancel(lastChannel.hashCode());
+            }
+        } catch (Throwable th) {
+            Log.e(getClass().getCanonicalName(), "Error when show notification", th);
         }
 
         // Check if running on a TV
@@ -271,6 +390,39 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                     mlastToast.show();
                 }
             });
+    }
+
+    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
+    private void lpNotificationExtras(Notification.Builder nbuilder, String category) {
+        nbuilder.setCategory(category);
+        nbuilder.setLocalOnly(true);
+
+    }
+
+    private boolean runningOnAndroidTV() {
+        UiModeManager uiModeManager = (UiModeManager) getSystemService(UI_MODE_SERVICE);
+        return uiModeManager.getCurrentModeType() == Configuration.UI_MODE_TYPE_TELEVISION;
+    }
+
+    @TargetApi(Build.VERSION_CODES.JELLY_BEAN)
+    private void jbNotificationExtras(int priority,
+                                      android.app.Notification.Builder nbuilder) {
+        try {
+            if (priority != 0) {
+                Method setpriority = nbuilder.getClass().getMethod("setPriority", int.class);
+                setpriority.invoke(nbuilder, priority);
+
+                Method setUsesChronometer = nbuilder.getClass().getMethod("setUsesChronometer", boolean.class);
+                setUsesChronometer.invoke(nbuilder, true);
+
+            }
+
+            //ignore exception
+        } catch (NoSuchMethodException | IllegalArgumentException |
+                InvocationTargetException | IllegalAccessException e) {
+            VpnStatus.logException(e);
+        }
+
     }
 
     @RequiresApi(api = Build.VERSION_CODES.JELLY_BEAN)
@@ -297,6 +449,16 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         }
     }
 
+    PendingIntent getUserInputIntent(String needed) {
+        Intent intent = new Intent(getApplicationContext(), LaunchVPN.class);
+        intent.addFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
+        intent.putExtra("need", needed);
+        Bundle b = new Bundle();
+        b.putString("need", needed);
+        PendingIntent pIntent = PendingIntent.getActivity(this, 12, intent, 0);
+        return pIntent;
+    }
+
     PendingIntent getGraphPendingIntent() {
         // Let the configure Button show the Log
 
@@ -312,71 +474,6 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
 
     }
 
-    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-    private void lpNotificationExtras(Notification.Builder nbuilder, String category) {
-        nbuilder.setCategory(category);
-        nbuilder.setLocalOnly(true);
-
-    }
-
-    private boolean runningOnAndroidTV() {
-        UiModeManager uiModeManager = (UiModeManager) getSystemService(UI_MODE_SERVICE);
-        return uiModeManager.getCurrentModeType() == Configuration.UI_MODE_TYPE_TELEVISION;
-    }
-
-    // setting notification status
-    private int getIconByConnectionStatus(ConnectionStatus level) {
-        switch (level) {
-            case LEVEL_CONNECTED:
-                return R.drawable.ic_stat_vpn;
-            case LEVEL_AUTH_FAILED:
-            case LEVEL_NONETWORK:
-            case LEVEL_NOTCONNECTED:
-                return R.drawable.ic_stat_vpn_offline;
-            case LEVEL_CONNECTING_NO_SERVER_REPLY_YET:
-            case LEVEL_WAITING_FOR_USER_INPUT:
-                return R.drawable.ic_stat_vpn_outline;
-            case LEVEL_CONNECTING_SERVER_REPLIED:
-                return R.drawable.ic_stat_vpn_empty_halo;
-            case LEVEL_VPNPAUSED:
-                return android.R.drawable.ic_media_pause;
-            case UNKNOWN_LEVEL:
-            default:
-                return R.drawable.ic_stat_vpn;
-        }
-    }
-
-    public static String getStatus() {//it will be call from mainactivity for get current status
-        return state;
-    }
-
-    @TargetApi(Build.VERSION_CODES.JELLY_BEAN)
-    private void jbNotificationExtras(int priority,
-                                      android.app.Notification.Builder nbuilder) {
-        try {
-            if (priority != 0) {
-                Method setpriority = nbuilder.getClass().getMethod("setPriority", int.class);
-                setpriority.invoke(nbuilder, priority);
-
-                Method setUsesChronometer = nbuilder.getClass().getMethod("setUsesChronometer", boolean.class);
-                setUsesChronometer.invoke(nbuilder, true);
-
-            }
-
-            //ignore exception
-        } catch (NoSuchMethodException | IllegalArgumentException |
-                InvocationTargetException | IllegalAccessException e) {
-            VpnStatus.logException(e);
-        }
-
-    }
-
-    PendingIntent getLogPendingIntent() {
-        Intent disconnectVPN = new Intent(this, DisconnectVPNActivity.class);
-        disconnectVPN.setAction(DISCONNECT_VPN);
-        return PendingIntent.getActivity(this, 0, disconnectVPN, 0);
-    }
-
     synchronized void registerDeviceStateReceiver(OpenVPNManagement magnagement) {
         // Registers BroadcastReceiver to track network connection changes.
         IntentFilter filter = new IntentFilter();
@@ -384,62 +481,112 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         filter.addAction(Intent.ACTION_SCREEN_OFF);
         filter.addAction(Intent.ACTION_SCREEN_ON);
         mDeviceStateReceiver = new DeviceStateReceiver(magnagement);
+
+        // Fetch initial network state
+        mDeviceStateReceiver.networkStateChange(this);
+
         registerReceiver(mDeviceStateReceiver, filter);
         VpnStatus.addByteCountListener(mDeviceStateReceiver);
+
         /*if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
             addLollipopCMListener(); */
     }
 
     synchronized void unregisterDeviceStateReceiver() {
-        if (mDeviceStateReceiver != null) try {
-            VpnStatus.removeByteCountListener(mDeviceStateReceiver);
-            this.unregisterReceiver(mDeviceStateReceiver);
-        } catch (IllegalArgumentException iae) {
-            // I don't know why  this happens:
-            // java.lang.IllegalArgumentException: Receiver not registered: de.blinkt.openvpn.NetworkSateReceiver@41a61a10
-            // Ignore for now ...
-            iae.printStackTrace();
-        }
+        if (mDeviceStateReceiver != null)
+            try {
+                VpnStatus.removeByteCountListener(mDeviceStateReceiver);
+                this.unregisterReceiver(mDeviceStateReceiver);
+            } catch (IllegalArgumentException iae) {
+                // I don't know why  this happens:
+                // java.lang.IllegalArgumentException: Receiver not registered: de.blinkt.openvpn.NetworkSateReceiver@41a61a10
+                // Ignore for now ...
+                iae.printStackTrace();
+            }
         mDeviceStateReceiver = null;
+
         /*if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
             removeLollipopCMListener();*/
+
     }
 
     public void userPause(boolean shouldBePaused) {
-        if (mDeviceStateReceiver != null) mDeviceStateReceiver.userPause(shouldBePaused);
+        if (mDeviceStateReceiver != null)
+            mDeviceStateReceiver.userPause(shouldBePaused);
+    }
+
+    @Override
+    public boolean stopVPN(boolean replaceConnection) throws RemoteException {
+        if (getManagement() != null)
+            return getManagement().stopVPN(replaceConnection);
+        else
+            return false;
     }
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
+
         if (intent != null && intent.getBooleanExtra(ALWAYS_SHOW_NOTIFICATION, false))
             mNotificationAlwaysVisible = true;
+
         VpnStatus.addStateListener(this);
         VpnStatus.addByteCountListener(this);
+
         guiHandler = new Handler(getMainLooper());
+
+        if (intent != null && DISCONNECT_VPN.equals(intent.getAction())) {
+            try {
+                stopVPN(false);
+            } catch (RemoteException e) {
+                VpnStatus.logException(e);
+            }
+            return START_NOT_STICKY;
+        }
+
         if (intent != null && PAUSE_VPN.equals(intent.getAction())) {
-            if (mDeviceStateReceiver != null) mDeviceStateReceiver.userPause(true);
+            if (mDeviceStateReceiver != null)
+                mDeviceStateReceiver.userPause(true);
             return START_NOT_STICKY;
         }
+
         if (intent != null && RESUME_VPN.equals(intent.getAction())) {
-            if (mDeviceStateReceiver != null) mDeviceStateReceiver.userPause(false);
+            if (mDeviceStateReceiver != null)
+                mDeviceStateReceiver.userPause(false);
             return START_NOT_STICKY;
         }
-        if (intent != null && START_SERVICE.equals(intent.getAction())) return START_NOT_STICKY;
+
+
+        if (intent != null && START_SERVICE.equals(intent.getAction()))
+            return START_NOT_STICKY;
         if (intent != null && START_SERVICE_STICKY.equals(intent.getAction())) {
             return START_REDELIVER_INTENT;
         }
+
+        // Always show notification here to avoid problem with startForeground timeout
+        VpnStatus.logInfo(R.string.building_configration);
+        VpnStatus.updateStateString("VPN_GENERATE_CONFIG", "", R.string.building_configration, ConnectionStatus.LEVEL_START);
+        showNotification(VpnStatus.getLastCleanLogMessage(this),
+                VpnStatus.getLastCleanLogMessage(this), NOTIFICATION_CHANNEL_NEWSTATUS_ID, 0, ConnectionStatus.LEVEL_START, null);
+
         if (intent != null && intent.hasExtra(getPackageName() + ".profileUUID")) {
             String profileUUID = intent.getStringExtra(getPackageName() + ".profileUUID");
-            mProfile = ProfileManager.get(this, profileUUID);
-            Log.d(TAG, "onStartCommand: " + mProfile);
+            int profileVersion = intent.getIntExtra(getPackageName() + ".profileVersion", 0);
+            // Try for 10s to get current version of the profile
+            mProfile = ProfileManager.get(this, profileUUID, profileVersion, 100);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1) {
+                updateShortCutUsage(mProfile);
+            }
+
         } else {
             /* The intent is null when we are set as always-on or the service has been restarted. */
             mProfile = ProfileManager.getLastConnectedProfile(this);
             VpnStatus.logInfo(R.string.service_restarted);
+
             /* Got no profile, just stop */
             if (mProfile == null) {
                 Log.d("OpenVPN", "Got no last connected profile on null intent. Assuming always on.");
                 mProfile = ProfileManager.getAlwaysOnVPN(this);
+
                 if (mProfile == null) {
                     stopSelf(startId);
                     return START_NOT_STICKY;
@@ -447,9 +594,14 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
             }
             /* Do the asynchronous keychain certificate stuff */
             mProfile.checkForRestart(this);
-            /* Recreate the intent */
-            intent = mProfile.getStartServiceIntent(this);
         }
+
+        if (mProfile == null) {
+            stopSelf(startId);
+            return START_NOT_STICKY;
+        }
+
+
         /* start the OpenVPN process itself in a background thread */
         new Thread(new Runnable() {
             @Override
@@ -457,16 +609,23 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                 startOpenVPN();
             }
         }).start();
+
+
         ProfileManager.setConnectedVpnProfile(this, mProfile);
-        /* TODO: At the moment we have no way to handle asynchronous PW input
-         * Fixing will also allow to handle challenge/response authentication */
-        if (mProfile.needUserPWInput(true) != 0) return START_NOT_STICKY;
+        VpnStatus.setConnectedVPNProfile(mProfile.getUUIDString());
+
         return START_STICKY;
     }
 
+    @RequiresApi(Build.VERSION_CODES.N_MR1)
+    private void updateShortCutUsage(VpnProfile profile) {
+        if (profile == null)
+            return;
+        ShortcutManager shortcutManager = getSystemService(ShortcutManager.class);
+        shortcutManager.reportShortcutUsed(profile.getUUIDString());
+    }
+
     private void startOpenVPN() {
-        VpnStatus.logInfo(R.string.building_configration);
-        VpnStatus.updateStateString("VPN_GENERATE_CONFIG", "", R.string.building_configration, VpnStatus.ConnectionStatus.LEVEL_START);
         try {
             mProfile.writeConfigFile(this);
         } catch (IOException e) {
@@ -474,24 +633,32 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
             endVpnService();
             return;
         }
-        // Extract information from the intent.
-        String prefix = getPackageName();
         String nativeLibraryDirectory = getApplicationInfo().nativeLibraryDir;
-        // Also writes OpenVPN binary
+        String tmpDir;
+        try {
+            tmpDir = getApplication().getCacheDir().getCanonicalPath();
+        } catch (IOException e) {
+            e.printStackTrace();
+            tmpDir = "/tmp";
+        }
+
+        // Write OpenVPN binary
         String[] argv = VPNLaunchHelper.buildOpenvpnArgv(this);
-        Log.d(TAG, "startOpenVPN: argv:" + (argv == null ? "null" : argv.length));
+
+
         // Set a flag that we are starting a new VPN
         mStarting = true;
         // Stop the previous session by interrupting the thread.
+
         stopOldOpenVPNProcess();
         // An old running VPN should now be exited
         mStarting = false;
+
         // Start a new session by creating a new thread.
-        SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(this);
-        mOvpn3 = prefs.getBoolean("ovpn3", false);
-//        if (!"ovpn3".equals(BuildConfig.FLAVOR)) mOvpn3 = false;
+        boolean useOpenVPN3 = VpnProfile.doUseOpenVPN3(this);
+
         // Open the Management Interface
-        if (!mOvpn3) {
+        if (!useOpenVPN3) {
             // start a Thread that handles incoming messages of the managment socket
             OpenVpnManagementThread ovpnManagementThread = new OpenVpnManagementThread(mProfile, this);
             if (ovpnManagementThread.openManagementInterface(this)) {
@@ -504,39 +671,50 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                 return;
             }
         }
+
         Runnable processThread;
-        if (mOvpn3) {
+        if (useOpenVPN3) {
             OpenVPNManagement mOpenVPN3 = instantiateOpenVPN3Core();
             processThread = (Runnable) mOpenVPN3;
             mManagement = mOpenVPN3;
         } else {
-            processThread = new OpenVPNThread(this, argv, nativeLibraryDirectory);
+            processThread = new OpenVPNThread(this, argv, nativeLibraryDirectory, tmpDir);
             mOpenVPNThread = processThread;
         }
+
         synchronized (mProcessLock) {
             mProcessThread = new Thread(processThread, "OpenVPNProcessThread");
             mProcessThread.start();
         }
+
         new Handler(getMainLooper()).post(new Runnable() {
-            @Override
-            public void run() {
-                if (mDeviceStateReceiver != null) unregisterDeviceStateReceiver();
-                registerDeviceStateReceiver(mManagement);
-            }
-        });
+                                              @Override
+                                              public void run() {
+                                                  if (mDeviceStateReceiver != null)
+                                                      unregisterDeviceStateReceiver();
+
+                                                  registerDeviceStateReceiver(mManagement);
+                                              }
+                                          }
+
+        );
     }
+
 
     private void stopOldOpenVPNProcess() {
         if (mManagement != null) {
-            if (mOpenVPNThread != null) ((OpenVPNThread) mOpenVPNThread).setReplaceConnection();
+            if (mOpenVPNThread != null)
+                ((OpenVPNThread) mOpenVPNThread).setReplaceConnection();
             if (mManagement.stopVPN(true)) {
                 // an old was asked to exit, wait 1s
                 try {
                     Thread.sleep(1000);
-                } catch (InterruptedException ignore) {
+                } catch (InterruptedException e) {
+                    //ignore
                 }
             }
         }
+
         forceStopOpenVpnProcess();
     }
 
@@ -546,7 +724,8 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                 mProcessThread.interrupt();
                 try {
                     Thread.sleep(1000);
-                } catch (InterruptedException ignore) {
+                } catch (InterruptedException e) {
+                    //ignore
                 }
             }
         }
@@ -563,6 +742,17 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         return null;
     }
 
+
+    @Override
+    public IBinder asBinder() {
+        return mBinder;
+    }
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+    }
+
     @Override
     public void onDestroy() {
         sendMessage("DISCONNECTED");
@@ -571,6 +761,7 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                 mManagement.stopVPN(true);
             }
         }
+
         if (mDeviceStateReceiver != null) {
             this.unregisterReceiver(mDeviceStateReceiver);
         }
@@ -583,8 +774,13 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         // The format of the string is not important, only that
         // two identical configurations produce the same result
         String cfg = "TUNCFG UNQIUE STRING ips:";
-        if (mLocalIP != null) cfg += mLocalIP.toString();
-        if (mLocalIPv6 != null) cfg += mLocalIPv6;
+
+        if (mLocalIP != null)
+            cfg += mLocalIP.toString();
+        if (mLocalIPv6 != null)
+            cfg += mLocalIPv6;
+
+
         cfg += "routes: " + TextUtils.join("|", mRoutes.getNetworks(true)) + TextUtils.join("|", mRoutesv6.getNetworks(true));
         cfg += "excl. routes:" + TextUtils.join("|", mRoutes.getNetworks(false)) + TextUtils.join("|", mRoutesv6.getNetworks(false));
         cfg += "dns: " + TextUtils.join("|", mDnslist);
@@ -594,18 +790,27 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
     }
 
     public ParcelFileDescriptor openTun() {
+
         //Debug.startMethodTracing(getExternalFilesDir(null).toString() + "/opentun.trace", 40* 1024 * 1024);
+
         Builder builder = new Builder();
+
         VpnStatus.logInfo(R.string.last_openvpn_tun_config);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && mProfile.mAllowLocalLAN) {
+
+        boolean allowUnsetAF = Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && !mProfile.mBlockUnusedAddressFamilies;
+        if (allowUnsetAF) {
             allowAllAFFamilies(builder);
         }
+
         if (mLocalIP == null && mLocalIPv6 == null) {
             VpnStatus.logError(getString(R.string.opentun_no_ipaddr));
             return null;
         }
+
         if (mLocalIP != null) {
-            addLocalNetworksToRoutes();
+            // OpenVPN3 manages excluded local networks by callback
+            if (!VpnProfile.doUseOpenVPN3(this))
+                addLocalNetworksToRoutes();
             try {
                 builder.addAddress(mLocalIP.mIp, mLocalIP.len);
             } catch (IllegalArgumentException iae) {
@@ -622,7 +827,10 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                 VpnStatus.logError(R.string.ip_add_error, mLocalIPv6, iae.getLocalizedMessage());
                 return null;
             }
+
         }
+
+
         for (String dns : mDnslist) {
             try {
                 builder.addDnsServer(dns);
@@ -630,21 +838,26 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                 VpnStatus.logError(R.string.dns_add_error, dns, iae.getLocalizedMessage());
             }
         }
+
         String release = Build.VERSION.RELEASE;
-        if ((Build.VERSION.SDK_INT == Build.VERSION_CODES.KITKAT && !release.startsWith("4.4.3") && !release.startsWith("4.4.4") && !release.startsWith("4.4.5") && !release.startsWith("4.4.6")) && mMtu < 1280) {
+        if ((Build.VERSION.SDK_INT == Build.VERSION_CODES.KITKAT && !release.startsWith("4.4.3")
+                && !release.startsWith("4.4.4") && !release.startsWith("4.4.5") && !release.startsWith("4.4.6"))
+                && mMtu < 1280) {
             VpnStatus.logInfo(String.format(Locale.US, "Forcing MTU to 1280 instead of %d to workaround Android Bug #70916", mMtu));
             builder.setMtu(1280);
         } else {
             builder.setMtu(mMtu);
         }
-        Collection<ipAddress> positiveIPv4Routes = mRoutes.getPositiveIPList();
-        Collection<ipAddress> positiveIPv6Routes = mRoutesv6.getPositiveIPList();
+
+        Collection<IpAddress> positiveIPv4Routes = mRoutes.getPositiveIPList();
+        Collection<IpAddress> positiveIPv6Routes = mRoutesv6.getPositiveIPList();
+
         if ("samsung".equals(Build.BRAND) && Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && mDnslist.size() >= 1) {
             // Check if the first DNS Server is in the VPN range
             try {
-                ipAddress dnsServer = new ipAddress(new CIDRIP(mDnslist.get(0), 32), true);
+                IpAddress dnsServer = new IpAddress(new CIDRIP(mDnslist.get(0), 32), true);
                 boolean dnsIncluded = false;
-                for (ipAddress net : positiveIPv4Routes) {
+                for (IpAddress net : positiveIPv4Routes) {
                     if (net.containsNet(dnsServer)) {
                         dnsIncluded = true;
                     }
@@ -655,28 +868,65 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                     positiveIPv4Routes.add(dnsServer);
                 }
             } catch (Exception e) {
-                VpnStatus.logError("Error parsing DNS Server IP: " + mDnslist.get(0));
+                // If it looks like IPv6 ignore error
+                if (!mDnslist.get(0).contains(":"))
+                    VpnStatus.logError("Error parsing DNS Server IP: " + mDnslist.get(0));
             }
         }
-        ipAddress multicastRange = new ipAddress(new CIDRIP("224.0.0.0", 3), true);
-        for (NetworkSpace.ipAddress route : positiveIPv4Routes) {
+
+        IpAddress multicastRange = new IpAddress(new CIDRIP("224.0.0.0", 3), true);
+
+        for (IpAddress route : positiveIPv4Routes) {
             try {
+
                 if (multicastRange.containsNet(route))
                     VpnStatus.logDebug(R.string.ignore_multicast_route, route.toString());
-                else builder.addRoute(route.getIPv4Address(), route.networkMask);
+                else
+                    builder.addRoute(route.getIPv4Address(), route.networkMask);
             } catch (IllegalArgumentException ia) {
                 VpnStatus.logError(getString(R.string.route_rejected) + route + " " + ia.getLocalizedMessage());
             }
         }
-        for (NetworkSpace.ipAddress route6 : positiveIPv6Routes) {
+
+        for (IpAddress route6 : positiveIPv6Routes) {
             try {
                 builder.addRoute(route6.getIPv6Address(), route6.networkMask);
             } catch (IllegalArgumentException ia) {
                 VpnStatus.logError(getString(R.string.route_rejected) + route6 + " " + ia.getLocalizedMessage());
             }
         }
+
+
+        if (mDomain != null)
+            builder.addSearchDomain(mDomain);
+
+        String ipv4info;
+        String ipv6info;
+        if (allowUnsetAF) {
+            ipv4info = "(not set, allowed)";
+            ipv6info = "(not set, allowed)";
+        } else {
+            ipv4info = "(not set)";
+            ipv6info = "(not set)";
+        }
+
+        int ipv4len;
+        if (mLocalIP != null) {
+            ipv4len = mLocalIP.len;
+            ipv4info = mLocalIP.mIp;
+        } else {
+            ipv4len = -1;
+        }
+
+        if (mLocalIPv6 != null) {
+            ipv6info = mLocalIPv6;
+        }
+
+        if ((!mRoutes.getNetworks(false).isEmpty() || !mRoutesv6.getNetworks(false).isEmpty()) && isLockdownEnabledCompat()) {
+            VpnStatus.logInfo("VPN lockdown enabled (do not allow apps to bypass VPN) enabled. Route exclusion will not allow apps to bypass VPN (e.g. bypass VPN for local networks)");
+        }
         if (mDomain != null) builder.addSearchDomain(mDomain);
-        VpnStatus.logInfo(R.string.local_ip_info, mLocalIP.mIp, mLocalIP.len, mLocalIPv6, mMtu);
+        VpnStatus.logInfo(R.string.local_ip_info, ipv4info, ipv4len, ipv6info, mMtu);
         VpnStatus.logInfo(R.string.dns_server_info, TextUtils.join(", ", mDnslist), mDomain);
         VpnStatus.logInfo(R.string.routes_info_incl, TextUtils.join(", ", mRoutes.getNetworks(true)), TextUtils.join(", ", mRoutesv6.getNetworks(true)));
         VpnStatus.logInfo(R.string.routes_info_excl, TextUtils.join(", ", mRoutes.getNetworks(false)), TextUtils.join(", ", mRoutesv6.getNetworks(false)));
@@ -684,15 +934,28 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
             setAllowedVpnPackages(builder);
         }
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP_MR1) {
+            // VPN always uses the default network
+            builder.setUnderlyingNetworks(null);
+        }
+
+
         String session = mProfile.mName;
         if (mLocalIP != null && mLocalIPv6 != null)
             session = getString(R.string.session_ipv6string, session, mLocalIP, mLocalIPv6);
         else if (mLocalIP != null)
             session = getString(R.string.session_ipv4string, session, mLocalIP);
+        else
+            session = getString(R.string.session_ipv4string, session, mLocalIPv6);
+
         builder.setSession(session);
+
         // No DNS Server, log a warning
-        if (mDnslist.size() == 0) VpnStatus.logInfo(R.string.warn_no_dns);
+        if (mDnslist.size() == 0)
+            VpnStatus.logInfo(R.string.warn_no_dns);
+
         mLastTunCfg = getTunConfigString();
+
         // Reset information
         mDnslist.clear();
         mRoutes.clear();
@@ -700,7 +963,9 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         mLocalIP = null;
         mLocalIPv6 = null;
         mDomain = null;
-        builder.setConfigureIntent(getLogPendingIntent());
+
+        builder.setConfigureIntent(getGraphPendingIntent());
+
         try {
             //Debug.stopMethodTracing();
             ParcelFileDescriptor tun = builder.establish();
@@ -715,6 +980,17 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
             }
             return null;
         }
+
+    }
+
+    private boolean isLockdownEnabledCompat() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            return isLockdownEnabled();
+        } else {
+            /* We cannot determine this, return false */
+            return false;
+        }
+
     }
 
     @TargetApi(Build.VERSION_CODES.LOLLIPOP)
@@ -724,43 +1000,70 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
     }
 
     private void addLocalNetworksToRoutes() {
-        // Add local network interfaces
-        String[] localRoutes = NativeUtils.getIfconfig();
-        // The format of mLocalRoutes is kind of broken because I don't really like JNI
-        for (int i = 0; i < localRoutes.length; i += 3) {
-            String intf = localRoutes[i];
-            String ipAddr = localRoutes[i + 1];
-            String netMask = localRoutes[i + 2];
-            if (intf == null || intf.equals("lo") ||
-                    intf.startsWith("tun") || intf.startsWith("rmnet")) continue;
-            if (ipAddr == null || netMask == null) {
-                VpnStatus.logError("Local routes are broken?! (Report to author) " + TextUtils.join("|", localRoutes));
+        for (String net : NetworkUtils.getLocalNetworks(this, false)) {
+            String[] netparts = net.split("/");
+            String ipAddr = netparts[0];
+            int netMask = Integer.parseInt(netparts[1]);
+            if (ipAddr.equals(mLocalIP.mIp))
                 continue;
-            }
-            if (ipAddr.equals(mLocalIP.mIp)) continue;
+
             if (Build.VERSION.SDK_INT < Build.VERSION_CODES.KITKAT && !mProfile.mAllowLocalLAN) {
                 mRoutes.addIPSplit(new CIDRIP(ipAddr, netMask), true);
+
             } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT && mProfile.mAllowLocalLAN)
                 mRoutes.addIP(new CIDRIP(ipAddr, netMask), false);
         }
+
+        // IPv6 is Lollipop+ only so we can skip the lower than KITKAT case
+        if (mProfile.mAllowLocalLAN) {
+            for (String net : NetworkUtils.getLocalNetworks(this, true)) {
+                addRoutev6(net, false);
+            }
+        }
+
+
     }
+
 
     @TargetApi(Build.VERSION_CODES.LOLLIPOP)
     private void setAllowedVpnPackages(Builder builder) {
+        boolean profileUsesOrBot = false;
+
+        for (Connection c : mProfile.mConnections) {
+            if (c.mProxyType == Connection.ProxyType.ORBOT)
+                profileUsesOrBot = true;
+        }
+
+        if (profileUsesOrBot)
+            VpnStatus.logDebug("VPN Profile uses at least one server entry with Orbot. Setting up VPN so that OrBot is not redirected over VPN.");
+
+
         boolean atLeastOneAllowedApp = false;
+
+        if (mProfile.mAllowedAppsVpnAreDisallowed && profileUsesOrBot) {
+            try {
+                builder.addDisallowedApplication(ORBOT_PACKAGE_NAME);
+            } catch (PackageManager.NameNotFoundException e) {
+                VpnStatus.logDebug("Orbot not installed?");
+            }
+        }
+
         for (String pkg : mProfile.mAllowedAppsVpn) {
             try {
                 if (mProfile.mAllowedAppsVpnAreDisallowed) {
                     builder.addDisallowedApplication(pkg);
                 } else {
-                    builder.addAllowedApplication(pkg);
-                    atLeastOneAllowedApp = true;
+                    if (!(profileUsesOrBot && pkg.equals(ORBOT_PACKAGE_NAME))) {
+                        builder.addAllowedApplication(pkg);
+                        atLeastOneAllowedApp = true;
+                    }
                 }
             } catch (PackageManager.NameNotFoundException e) {
                 mProfile.mAllowedAppsVpn.remove(pkg);
                 VpnStatus.logInfo(R.string.app_no_longer_exists, pkg);
             }
         }
+
         if (!mProfile.mAllowedAppsVpnAreDisallowed && !atLeastOneAllowedApp) {
             VpnStatus.logDebug(R.string.no_allowed_app, getPackageName());
             try {
@@ -769,10 +1072,16 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                 VpnStatus.logError("This should not happen: " + e.getLocalizedMessage());
             }
         }
+
         if (mProfile.mAllowedAppsVpnAreDisallowed) {
             VpnStatus.logDebug(R.string.disallowed_vpn_apps_info, TextUtils.join(", ", mProfile.mAllowedAppsVpn));
         } else {
             VpnStatus.logDebug(R.string.allowed_vpn_apps_info, TextUtils.join(", ", mProfile.mAllowedAppsVpn));
+        }
+
+        if (mProfile.mAllowAppVpnBypass) {
+            builder.allowBypass();
+            VpnStatus.logDebug("Apps may bypass VPN");
         }
     }
 
@@ -789,45 +1098,63 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
     /**
      * Route that is always included, used by the v3 core
      */
-    public void addRoute(CIDRIP route) {
-        mRoutes.addIP(route, true);
+    public void addRoute(CIDRIP route, boolean include) {
+        mRoutes.addIP(route, include);
     }
 
     public void addRoute(String dest, String mask, String gateway, String device) {
         CIDRIP route = new CIDRIP(dest, mask);
         boolean include = isAndroidTunDevice(device);
-        NetworkSpace.ipAddress gatewayIP = new NetworkSpace.ipAddress(new CIDRIP(gateway, 32), false);
+
+        IpAddress gatewayIP = new IpAddress(new CIDRIP(gateway, 32), false);
+
         if (mLocalIP == null) {
             VpnStatus.logError("Local IP address unset and received. Neither pushed server config nor local config specifies an IP addresses. Opening tun device is most likely going to fail.");
             return;
         }
-        NetworkSpace.ipAddress localNet = new NetworkSpace.ipAddress(mLocalIP, true);
-        if (localNet.containsNet(gatewayIP)) include = true;
-        if (gateway != null && (gateway.equals("255.255.255.255") || gateway.equals(mRemoteGW)))
+        IpAddress localNet = new IpAddress(mLocalIP, true);
+        if (localNet.containsNet(gatewayIP))
             include = true;
+
+        if (gateway != null &&
+                (gateway.equals("255.255.255.255") || gateway.equals(mRemoteGW)))
+            include = true;
+
+
         if (route.len == 32 && !mask.equals("255.255.255.255")) {
             VpnStatus.logWarning(R.string.route_not_cidr, dest, mask);
         }
+
         if (route.normalise())
             VpnStatus.logWarning(R.string.route_not_netip, dest, route.len, route.mIp);
+
         mRoutes.addIP(route, include);
     }
 
     public void addRoutev6(String network, String device) {
-        String[] v6parts = network.split("/");
-        boolean included = isAndroidTunDevice(device);
         // Tun is opened after ROUTE6, no device name may be present
+        boolean included = isAndroidTunDevice(device);
+        addRoutev6(network, included);
+    }
+
+    public void addRoutev6(String network, boolean included) {
+        String[] v6parts = network.split("/");
+
         try {
             Inet6Address ip = (Inet6Address) InetAddress.getAllByName(v6parts[0])[0];
             int mask = Integer.parseInt(v6parts[1]);
             mRoutesv6.addIPv6(ip, mask, included);
+
         } catch (UnknownHostException e) {
             VpnStatus.logException(e);
         }
+
+
     }
 
     private boolean isAndroidTunDevice(String device) {
-        return device != null && (device.startsWith("tun") || "(null)".equals(device) || "vpnservice-tun".equals(device));
+        return device != null &&
+                (device.startsWith("tun") || "(null)".equals(device) || VPNSERVICE_TUN.equals(device));
     }
 
     public void setMtu(int mtu) {
@@ -842,9 +1169,12 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         mLocalIP = new CIDRIP(local, netmask);
         mMtu = mtu;
         mRemoteGW = null;
+
         long netMaskAsInt = CIDRIP.getInt(netmask);
+
         if (mLocalIP.len == 32 && !netmask.equals("255.255.255.255")) {
             // get the netmask as IP
+
             int masklen;
             long mask;
             if ("net30".equals(mode)) {
@@ -854,6 +1184,7 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
                 masklen = 31;
                 mask = 0xfffffffe;
             }
+
             // Netmask is Ip address +/-1, assume net30/p2p with small net
             if ((netMaskAsInt & mask) == (mLocalIP.getInt() & mask)) {
                 mLocalIP.len = masklen;
@@ -866,12 +1197,16 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         if (("p2p".equals(mode) && mLocalIP.len < 32) || ("net30".equals(mode) && mLocalIP.len < 30)) {
             VpnStatus.logWarning(R.string.ip_looks_like_subnet, local, netmask, mode);
         }
+
+
         /* Workaround for Lollipop, it  does not route traffic to the VPNs own network mask */
         if (mLocalIP.len <= 31 && Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
             CIDRIP interfaceRoute = new CIDRIP(mLocalIP.mIp, mLocalIP.len);
             interfaceRoute.normalise();
-            addRoute(interfaceRoute);
+            addRoute(interfaceRoute, true);
         }
+
+
         // Configurations are sometimes really broken...
         mRemoteGW = netmask;
     }
@@ -881,49 +1216,50 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
     }
 
     @Override
-    public void updateState(String state, String logmessage, int resid, ConnectionStatus level) {
+    public void updateState(String state, String logmessage, int resid, ConnectionStatus level, Intent intent) {
         // If the process is not running, ignore any state,
         // Notification should be invisible in this state
+
         doSendBroadcast(state, level);
-        if (mProcessThread == null && !mNotificationAlwaysVisible) return;
-        boolean lowpriority = false;
+        if (mProcessThread == null && !mNotificationAlwaysVisible)
+            return;
+
         String channel = NOTIFICATION_CHANNEL_NEWSTATUS_ID;
-
         // Display byte count only after being connected
+
         {
-
-            if (level == LEVEL_WAITING_FOR_USER_INPUT) {
-                // The user is presented a dialog of some kind, no need to inform the user
-                // with a notifcation
-                return;
-            } else if (level == LEVEL_CONNECTED) {
-
+            if (level == LEVEL_CONNECTED) {
                 mDisplayBytecount = true;
                 mConnecttime = System.currentTimeMillis();
-
-                if (!runningOnAndroidTV()) lowpriority = true;
+                if (!runningOnAndroidTV())
+                    channel = NOTIFICATION_CHANNEL_BG_ID;
             } else {
                 mDisplayBytecount = false;
             }
+
             // Other notifications are shown,
             // This also mean we are no longer connected, ignore bytecount messages until next
             // CONNECTED
             // Does not work :(
             String msg = getString(resid);
             showNotification(VpnStatus.getLastCleanLogMessage(this),
-                    VpnStatus.getLastCleanLogMessage(this), channel, 0, level, null);
+                    VpnStatus.getLastCleanLogMessage(this), channel, 0, level, intent);
 
         }
     }
 
+    @Override
+    public void setConnectedVPN(String uuid) {
+    }
+
     private void doSendBroadcast(String state, ConnectionStatus level) {
         Intent vpnstatus = new Intent();
+        vpnstatus.setAction("de.blinkt.openvpn.VPN_STATUS");
         vpnstatus.putExtra("status", level.toString());
         vpnstatus.putExtra("detailstatus", state);
         sendBroadcast(vpnstatus, permission.ACCESS_NETWORK_STATE);
         sendMessage(state);
     }
-
 
     long c = Calendar.getInstance().getTimeInMillis();
     long time;
@@ -932,35 +1268,30 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
 
     @Override
     public void updateByteCount(long in, long out, long diffIn, long diffOut) {
+        TotalTraffic.calcTraffic(this, in, out, diffIn, diffOut);
         if (mDisplayBytecount) {
-
             String netstat = String.format(getString(R.string.statusline_bytecount),
-                    humanReadableByteCount(in, false),
-                    humanReadableByteCount(diffIn / OpenVPNManagement.mBytecountInterval, false),
-                    humanReadableByteCount(out, false),
-                    humanReadableByteCount(diffOut / OpenVPNManagement.mBytecountInterval, false));
+                    humanReadableByteCount(in, false, getResources()),
+                    humanReadableByteCount(diffIn / OpenVPNManagement.mBytecountInterval, true, getResources()),
+                    humanReadableByteCount(out, false, getResources()),
+                    humanReadableByteCount(diffOut / OpenVPNManagement.mBytecountInterval, true, getResources()));
 
-            boolean lowpriority = !mNotificationAlwaysVisible;
+
             showNotification(netstat, null, NOTIFICATION_CHANNEL_BG_ID, mConnecttime, LEVEL_CONNECTED, null);
-
-
             byteIn = String.format("↓%2$s", getString(R.string.statusline_bytecount),
-                    humanReadableByteCount(in, false)) + " - " + humanReadableByteCount(diffIn / OpenVPNManagement.mBytecountInterval, false) + "/s";
-
+                    humanReadableByteCount(in,false, getResources())) + " - " + humanReadableByteCount(diffIn / OpenVPNManagement.mBytecountInterval, false, getResources()) + "/s";
             byteOut = String.format("↑%2$s", getString(R.string.statusline_bytecount),
-                    humanReadableByteCount(out, false)) + " - " + humanReadableByteCount(diffOut / OpenVPNManagement.mBytecountInterval, false) + "/s";
-
+                    humanReadableByteCount(out, false,getResources())) + " - " + humanReadableByteCount(diffOut / OpenVPNManagement.mBytecountInterval, false, getResources()) + "/s";
             time = Calendar.getInstance().getTimeInMillis() - c;
             lastPacketReceive = Integer.parseInt(convertTwoDigit((int) (time / 1000) % 60)) - Integer.parseInt(seconds);
-
             seconds = convertTwoDigit((int) (time / 1000) % 60);
             minutes = convertTwoDigit((int) ((time / (1000 * 60)) % 60));
             hours = convertTwoDigit((int) ((time / (1000 * 60 * 60)) % 24));
             duration = hours + ":" + minutes + ":" + seconds;
             lastPacketReceive = checkPacketReceive(lastPacketReceive);
-
             sendMessage(duration, String.valueOf(lastPacketReceive), byteIn, byteOut);
         }
+
     }
 
     public int checkPacketReceive(int value) {
@@ -968,29 +1299,9 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
         if (value < 0) return 0;
         else return value;
     }
-
     public String convertTwoDigit(int value) {
         if (value < 10) return "0" + value;
         else return value + "";
-    }
-
-    //sending message to main activity
-    private void sendMessage(String state) {
-        Intent intent = new Intent("connectionState");
-        intent.putExtra("state", state);
-        this.state = state;
-        LocalBroadcastManager.getInstance(getApplicationContext()).sendBroadcast(intent);
-    }
-
-    //sending message to main activity
-    private void sendMessage(String duration, String lastPacketReceive, String byteIn, String byteOut) {
-        Intent intent = new Intent("connectionState");
-        intent.putExtra("duration", duration);
-        intent.putExtra("lastPacketReceive", lastPacketReceive);
-        intent.putExtra("byteIn", byteIn);
-        intent.putExtra("byteOut", byteOut);
-
-        LocalBroadcastManager.getInstance(getApplicationContext()).sendBroadcast(intent);
     }
 
     @Override
@@ -1014,24 +1325,107 @@ public class OpenVPNService extends VpnService implements StateListener, Callbac
             return "NOACTION";
         } else {
             String release = Build.VERSION.RELEASE;
-            if (Build.VERSION.SDK_INT == Build.VERSION_CODES.KITKAT && !release.startsWith("4.4.3") && !release.startsWith("4.4.4") && !release.startsWith("4.4.5") && !release.startsWith("4.4.6"))
+            if (Build.VERSION.SDK_INT == Build.VERSION_CODES.KITKAT && !release.startsWith("4.4.3")
+                    && !release.startsWith("4.4.4") && !release.startsWith("4.4.5") && !release.startsWith("4.4.6"))
                 // There will be probably no 4.4.4 or 4.4.5 version, so don't waste effort to do parsing here
                 return "OPEN_AFTER_CLOSE";
-            else return "OPEN_BEFORE_CLOSE";
+            else
+                return "OPEN_BEFORE_CLOSE";
         }
     }
 
+    public void requestInputFromUser(int resid, String needed) {
+        VpnStatus.updateStateString("NEED", "need " + needed, resid, LEVEL_WAITING_FOR_USER_INPUT);
+        showNotification(getString(resid), getString(resid), NOTIFICATION_CHANNEL_NEWSTATUS_ID, 0, LEVEL_WAITING_FOR_USER_INPUT, null);
+    }
+
+
+    public void trigger_sso(String info) {
+        String channel = NOTIFICATION_CHANNEL_USERREQ_ID;
+        String method = info.split(":", 2)[0];
+
+        NotificationManager mNotificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+
+        Notification.Builder nbuilder = new Notification.Builder(this);
+        nbuilder.setAutoCancel(true);
+        int icon = android.R.drawable.ic_dialog_info;
+        nbuilder.setSmallIcon(icon);
+
+        Intent intent;
+
+        int reason;
+        if (method.equals("CR_TEXT")) {
+            String challenge = info.split(":", 2)[1];
+            reason = R.string.crtext_requested;
+            nbuilder.setContentTitle(getString(reason));
+            nbuilder.setContentText(challenge);
+
+            intent = new Intent();
+            intent.setComponent(new ComponentName(this, getPackageName() + ".activities.CredentialsPopup"));
+
+            intent.putExtra(EXTRA_CHALLENGE_TXT, challenge);
+
+        } else {
+            VpnStatus.logError("Unknown SSO method found: " + method);
+            return;
+        }
+
+        // updateStateString trigger the notification of the VPN to be refreshed, save this intent
+        // to have that notification also this intent to be set
+        PendingIntent pIntent = PendingIntent.getActivity(this, 0, intent, 0);
+        VpnStatus.updateStateString("USER_INPUT", "waiting for user input", reason, LEVEL_WAITING_FOR_USER_INPUT, intent);
+        nbuilder.setContentIntent(pIntent);
+
+
+        // Try to set the priority available since API 16 (Jellybean)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN)
+            jbNotificationExtras(PRIORITY_MAX, nbuilder);
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
+            lpNotificationExtras(nbuilder, Notification.CATEGORY_STATUS);
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            //noinspection NewApi
+            nbuilder.setChannelId(channel);
+        }
+
+        @SuppressWarnings("deprecation")
+        Notification notification = nbuilder.getNotification();
+
+
+        int notificationId = channel.hashCode();
+
+        mNotificationManager.notify(notificationId, notification);
+    }
+
+    //sending message to main activity
+    private void sendMessage(String state) {
+        Intent intent = new Intent("connectionState");
+        intent.putExtra("state", state);
+        this.state = state;
+        LocalBroadcastManager.getInstance(getApplicationContext()).sendBroadcast(intent);
+    }
+    //sending message to main activity
+    private void sendMessage(String duration, String lastPacketReceive, String byteIn, String byteOut) {
+        Intent intent = new Intent("connectionState");
+        intent.putExtra("duration", duration);
+        intent.putExtra("lastPacketReceive", lastPacketReceive);
+        intent.putExtra("byteIn", byteIn);
+        intent.putExtra("byteOut", byteOut);
+        LocalBroadcastManager.getInstance(getApplicationContext()).sendBroadcast(intent);
+    }
     public class LocalBinder extends Binder {
         public OpenVPNService getService() {
             // Return this instance of LocalService so clients can call public methods
             return OpenVPNService.this;
         }
     }
-
+    public static String getStatus() {//it will be call from mainactivity for get current status
+        return state;
+    }
     public static void setDefaultStatus() {
         state = "";
     }
-
     public boolean isConnected() {
         return flag;
     }
