@@ -2,6 +2,7 @@
  * Copyright (c) 2012-2016 Arne Schwabe
  * Distributed under the GNU GPL v2 with additional terms. For full terms see the file doc/LICENSE.txt
  */
+
 package de.blinkt.openvpn.core;
 
 import android.annotation.SuppressLint;
@@ -22,59 +23,61 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import de.blinkt.openvpn.R;
-import de.blinkt.openvpn.core.VpnStatus.ConnectionStatus;
 
 public class OpenVPNThread implements Runnable {
-    public static final int M_FATAL = (1 << 4);
-    public static final int M_NONFATAL = (1 << 5);
-    public static final int M_WARN = (1 << 6);
-    public static final int M_DEBUG = (1 << 7);
     private static final String DUMP_PATH_STRING = "Dump path: ";
     @SuppressLint("SdCardPath")
     private static final String BROKEN_PIE_SUPPORT = "/data/data/de.blinkt.openvpn/cache/pievpn";
     private final static String BROKEN_PIE_SUPPORT2 = "syntax error";
-    private static final String TAG = "CakeVpn";
+    private static final String TAG = "OpenVPN";
+    // 1380308330.240114 18000002 Send to HTTP proxy: 'X-Online-Host: bla.blabla.com'
+    private static final Pattern LOG_PATTERN = Pattern.compile("(\\d+).(\\d+) ([0-9a-f])+ (.*)");
+    public static final int M_FATAL = (1 << 4);
+    public static final int M_NONFATAL = (1 << 5);
+    public static final int M_WARN = (1 << 6);
+    public static final int M_DEBUG = (1 << 7);
     private String[] mArgv;
     private static Process mProcess;
     private String mNativeDir;
+    private String mTmpDir;
     private static OpenVPNService mService;
     private String mDumpPath;
     private boolean mBrokenPie = false;
     private boolean mNoProcessExitStatus = false;
 
-    public static boolean flag = false;
-    public OpenVPNThread(OpenVPNService service, String[] argv, String nativelibdir) {
+    public OpenVPNThread(OpenVPNService service, String[] argv, String nativelibdir, String tmpdir) {
         mArgv = argv;
         mNativeDir = nativelibdir;
+        mTmpDir = tmpdir;
         mService = service;
     }
 
     public OpenVPNThread() {
-
     }
 
     public void stopProcess() {
         mProcess.destroy();
     }
 
-    void setReplaceConnection() {
-        mNoProcessExitStatus = true;
+    void setReplaceConnection()
+    {
+        mNoProcessExitStatus=true;
     }
 
     @Override
     public void run() {
         try {
-
-            flag = true;
+            Log.i(TAG, "Starting openvpn");
             startOpenVPNThreadArgs(mArgv);
-            Log.i(TAG, "VPN process exited");
+            Log.i(TAG, "OpenVPN process exited");
         } catch (Exception e) {
-            VpnStatus.logException("Starting Vpn Thread", e);
-            Log.e(TAG, "VPN Got " + e.toString());
+            VpnStatus.logException("Starting OpenVPN Thread", e);
+            Log.e(TAG, "OpenVPNThread Got " + e.toString());
         } finally {
             int exitvalue = 0;
             try {
-                if (mProcess != null) exitvalue = mProcess.waitFor();
+                if (mProcess != null)
+                    exitvalue = mProcess.waitFor();
             } catch (IllegalThreadStateException ite) {
                 VpnStatus.logError("Illegal Thread state: " + ite.getLocalizedMessage());
             } catch (InterruptedException ie) {
@@ -85,24 +88,28 @@ public class OpenVPNThread implements Runnable {
                 if (mBrokenPie) {
                     /* This will probably fail since the NoPIE binary is probably not written */
                     String[] noPieArgv = VPNLaunchHelper.replacePieWithNoPie(mArgv);
+
                     // We are already noPIE, nothing to gain
                     if (!noPieArgv.equals(mArgv)) {
                         mArgv = noPieArgv;
                         VpnStatus.logInfo("PIE Version could not be executed. Trying no PIE version");
                         run();
-                        return;
                     }
+
                 }
+
             }
-            if (!mNoProcessExitStatus) VpnStatus.updateStateString("NOPROCESS", "No process running.", R.string.state_noprocess, ConnectionStatus.LEVEL_NOTCONNECTED);
+
+            if (!mNoProcessExitStatus)
+                VpnStatus.updateStateString("NOPROCESS", "No process running.", R.string.state_noprocess, ConnectionStatus.LEVEL_NOTCONNECTED);
+
             if (mDumpPath != null) {
                 try {
                     BufferedWriter logout = new BufferedWriter(new FileWriter(mDumpPath + ".log"));
-                    SimpleDateFormat timeformat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+                    SimpleDateFormat timeformat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss", Locale.GERMAN);
                     for (LogItem li : VpnStatus.getlogbuffer()) {
                         String time = timeformat.format(new Date(li.getLogtime()));
                         logout.write(time + " " + li.getString(mService) + "\n");
-
                     }
                     logout.close();
                     VpnStatus.logError(R.string.minidump_generated);
@@ -110,18 +117,18 @@ public class OpenVPNThread implements Runnable {
                     VpnStatus.logError("Writing minidump log: " + e.getLocalizedMessage());
                 }
             }
-            mService.processDied();
+
+            if (!mNoProcessExitStatus)
+                mService.openvpnStopped();
             Log.i(TAG, "Exiting");
         }
     }
 
     public static boolean stop(){
-
-        mService.processDied();
+        mService.openvpnStopped();
         mProcess.destroy();
         return true;
     }
-
 
     private void startOpenVPNThreadArgs(String[] argv) {
         LinkedList<String> argvlist = new LinkedList<String>();
@@ -134,6 +141,7 @@ public class OpenVPNThread implements Runnable {
         String lbpath = genLibraryPath(argv, pb);
 
         pb.environment().put("LD_LIBRARY_PATH", lbpath);
+        pb.environment().put("TMPDIR", mTmpDir);
 
         pb.redirectErrorStream(true);
         try {
@@ -154,11 +162,7 @@ public class OpenVPNThread implements Runnable {
                 if (logline.startsWith(BROKEN_PIE_SUPPORT) || logline.contains(BROKEN_PIE_SUPPORT2))
                     mBrokenPie = true;
 
-
-                // 1380308330.240114 18000002 Send to HTTP proxy: 'X-Online-Host: bla.blabla.com'
-
-                Pattern p = Pattern.compile("(\\d+).(\\d+) ([0-9a-f])+ (.*)");
-                Matcher m = p.matcher(logline);
+                Matcher m = LOG_PATTERN.matcher(logline);
                 int logerror = 0;
                 if (m.matches()) {
                     int flags = Integer.parseInt(m.group(3), 16);
@@ -217,6 +221,4 @@ public class OpenVPNThread implements Runnable {
         }
         return lbpath;
     }
-
-
 }
